Function fastsim_core::cycle::detect_passing

pub fn detect_passing(
    cyc: &RustCycle,
    cyc0: &RustCycle,
    i: usize,
    dist_tol_m: Option<f64>,
) -> PassingInfo

Reports back information of the first point where cyc passes cyc0, starting at step i until the next stop of cyc.

cyc: fastsim.Cycle, the proposed cycle of the vehicle under simulation

cyc0: fastsim.Cycle, the reference/lead vehicle/shadow cycle to compare with

i: int, the time-step index to consider

dist_tol_m: float, the distance tolerance away from lead vehicle to be seen as “deviated” from the reference/shadow trace (m) RETURNS: PassingInfo