Struct fastsim_core::cycle::RustCycle

pub struct RustCycle {
    pub time_s: Array1<f64>,
    pub mps: Array1<f64>,
    pub grade: Array1<f64>,
    pub road_type: Array1<f64>,
    pub name: String,
    pub orphaned: bool,
}

Struct for containing:

time_s, cycle time, $s$

mps, vehicle speed, $\frac{m}{s}$

grade, road grade/slope, $\frac{rise}{run}$

road_type, $kW$

legacy, will likely change to road charging capacity

Another sublist.

Fields
time_s: Array1<f64>

array of time s

mps: Array1<f64>

array of speed [m/s]

grade: Array1<f64>

array of grade [rise/run]

road_type: Array1<f64>

array of max possible charge rate from roadway

name: Stringorphaned: bool
Implementations

impl RustCycle

pub fn __len__(&self) -> usize

pub fn __getnewargs__(&self) -> (Vec<f64>, Vec<f64>, Vec<f64>, Vec<f64>, &str)

pub fn from_csv_py(filepath: &PyAny, skip_init: Option<bool>) -> Result<Self>

pub fn to_rust(&self) -> Self

pub fn from_dict(dict: &PyDict, skip_init: Option<bool>) -> Result<Self>

pub fn to_dict<'py>(&self, py: Python<'py>) -> Result<&'py PyDict>

pub fn to_csv_py(&self) -> PyResult<String>

pub fn modify_by_const_jerk_trajectory_py( &mut self, idx: usize, n: usize, jerk_m_per_s3: f64, accel0_m_per_s2: f64, ) -> f64

pub fn modify_with_braking_trajectory_py( &mut self, brake_accel_m_per_s2: f64, idx: usize, dts_m: Option<f64>, ) -> Result<(f64, usize)>

pub fn calc_distance_to_next_stop_from_py(&self, distance_m: f64) -> f64

pub fn average_grade_over_range_py( &self, distance_start_m: f64, delta_distance_m: f64, ) -> f64

pub fn build_cache_py(&self) -> RustCycleCache

pub fn dt_s_at_i_py(&self, i: usize) -> f64

pub fn get_mph(&self) -> Vec<f64>

pub fn set_mph(&mut self, new_value: Vec<f64>)

pub fn get_dt_s(&self) -> Vec<f64>

array of time steps

pub fn get_dist_m(&self) -> Vec<f64>

distance for each time step based on final speed

pub fn get_delta_elev_m(&self) -> Vec<f64>

pub fn list_resources_py(&self) -> Vec<String>

list available cycle resources

pub fn get_time_s(&self) -> Pyo3ArrayF64

pub fn set_time_s(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mps(&self) -> Pyo3ArrayF64

pub fn set_mps(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_grade(&self) -> Pyo3ArrayF64

pub fn set_grade(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_road_type(&self) -> Pyo3ArrayF64

pub fn set_road_type(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_name(&self) -> String

pub fn set_name(&mut self, new_value: String) -> Result<()>

pub fn get_orphaned(&self) -> bool

pub fn reset_orphaned(&mut self)

Reset the orphaned flag to false.

pub fn copy(&self) -> Self

pub fn __copy__(&self) -> Self

pub fn __deepcopy__(&self, _memo: &PyDict) -> Self

pub fn from_resource_py( filepath: &PyAny, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a resource file packaged with the fastsim-core crate

Arguments:

filepath: str | pathlib.Path - Filepath, relative to the top of the resources folder (excluding any relevant prefix), from which to read the object

pub fn to_file_py(&self, filepath: &PyAny) -> PyResult<()>

Write (serialize) an object to a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS. Creates a new file if it does not already exist, otherwise truncates the existing file.

Arguments

filepath: str | pathlib.Path - The filepath at which to write the object

pub fn from_file_py(filepath: &PyAny, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS.

Arguments:

filepath: str | pathlib.Path - The filepath from which to read the object

pub fn to_str_py(&self, format: &str) -> PyResult<String>

Write (serialize) an object into a string

Arguments:

format: str - The target format, any of those listed in ACCEPTED_STR_FORMATS

pub fn from_str_py( contents: &str, format: &str, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a string

Arguments:

contents: str - The string containing the object data

format: str - The source format, any of those listed in ACCEPTED_STR_FORMATS

pub fn to_json_py(&self) -> PyResult<String>

Write (serialize) an object to a JSON string

pub fn from_json_py(json_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object to a JSON string

Arguments

json_str: str - JSON-formatted string to deserialize from

pub fn to_yaml_py(&self) -> PyResult<String>

Write (serialize) an object to a YAML string

pub fn from_yaml_py(yaml_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a YAML string

Arguments

yaml_str: str - YAML-formatted string to deserialize from

impl RustCycle

pure Rust methods that need to be separate due to pymethods incompatibility

pub fn from_csv_file<P: AsRef<Path>>( filepath: P, skip_init: bool, ) -> Result<Self>

Load cycle from CSV file, parsing name from filepath

pub fn from_csv_str<S: AsRef<str>>( csv_str: S, name: String, skip_init: bool, ) -> Result<Self>

Load cycle from CSV string

pub fn to_csv(&self) -> Result<String>

Write (serialize) cycle to a CSV string

pub fn build_cache(&self) -> RustCycleCache

pub fn push(&mut self, cyc_elem: RustCycleElement)

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn is_sorted(&self) -> bool

pub fn are_fields_equal_length(&self) -> bool

pub fn test_cyc() -> Self

pub fn average_grade_over_range( &self, distance_start_m: f64, delta_distance_m: f64, cache: Option<&RustCycleCache>, ) -> f64

Returns the average grade over the given range of distances

distance_start_m: non-negative-number, the distance at start of evaluation area (m)

delta_distance_m: non-negative-number, the distance traveled from distance_start_m (m) RETURN: number, the average grade (rise over run) over the given distance range Note: grade is assumed to be constant from just after the previous sample point until the current sample point. That is, grade[i] applies over the range of distances, d, from (d[i - 1], d[i])

pub fn calc_distance_to_next_stop_from( &self, distance_m: f64, cache: Option<&RustCycleCache>, ) -> f64

Calculate the distance to next stop from distance_m

distance_m: non-negative-number, the current distance from start (m) RETURN: returns the distance to the next stop from distance_m NOTE: distance may be negative if we’re beyond the last stop

pub fn modify_by_const_jerk_trajectory( &mut self, i: usize, n: usize, jerk_m_per_s3: f64, accel0_m_per_s2: f64, ) -> f64

Modifies the cycle using the given constant-jerk trajectory parameters

idx: non-negative integer, the point in the cycle to initiate modification (note: THIS point is modified since trajectory should be calculated from idx-1)

n: non-negative integer, the number of steps ahead

jerk_m__s3: number, the “Jerk” associated with the trajectory (m/s3)

accel0_m__s2: number, the initial acceleration (m/s2) NOTE:

modifies cyc in place to hit any critical rendezvous_points by a trajectory adjustment

CAUTION: NOT ROBUST AGAINST VARIABLE DURATION TIME-STEPS RETURN: Number, final modified speed (m/s)

pub fn modify_with_braking_trajectory( &mut self, brake_accel_m_per_s2: f64, i: usize, dts_m: Option<f64>, ) -> Result<(f64, usize)>

Add a braking trajectory that would cover the same distance as the given constant brake deceleration

brake_accel_m__s2: negative number, the braking acceleration (m/s2)

idx: non-negative integer, the index where to initiate the stop trajectory, start of the step (i in FASTSim)

dts_m: None | float: if given, this is the desired distance-to-stop in meters. If not given, it is calculated based on braking deceleration. RETURN: (non-negative-number, positive-integer)

the final speed of the modified trajectory (m/s)

the number of time-steps required to complete the braking maneuver NOTE:

modifies the cycle in place for the braking trajectory

pub fn dt_s(&self) -> Array1<f64>

rust-internal time steps

pub fn dt_s_at_i(&self, i: usize) -> f64

rust-internal time steps at i

pub fn dist_m(&self) -> Array1<f64>

distance covered in each time step

pub fn mph_at_i(&self, i: usize) -> f64

get mph from self.mps

pub fn delta_elev_m(&self) -> Array1<f64>

elevation change w.r.t. to initial