Struct fastsim_core::cycle::RustCycleCache

pub struct RustCycleCache {
    pub grade_all_zero: bool,
    pub trapz_step_distances_m: Array1<f64>,
    pub trapz_distances_m: Array1<f64>,
    pub trapz_elevations_m: Array1<f64>,
    pub stops: Array1<bool>,
    /* private fields */
}

Fields
grade_all_zero: booltrapz_step_distances_m: Array1<f64>trapz_distances_m: Array1<f64>trapz_elevations_m: Array1<f64>stops: Array1<bool>
Implementations

impl RustCycleCache

pub fn __new__(cyc: &RustCycle) -> Self

pub fn get_grade_all_zero(&self) -> bool

pub fn set_grade_all_zero(&mut self, new_value: bool) -> Result<()>

pub fn get_trapz_step_distances_m(&self) -> Pyo3ArrayF64

pub fn set_trapz_step_distances_m(&mut self, new_value: Vec<f64>)

pub fn get_trapz_distances_m(&self) -> Pyo3ArrayF64

pub fn set_trapz_distances_m(&mut self, new_value: Vec<f64>)

pub fn get_trapz_elevations_m(&self) -> Pyo3ArrayF64

pub fn set_trapz_elevations_m(&mut self, new_value: Vec<f64>)

pub fn get_stops(&self) -> Pyo3ArrayBool

pub fn set_stops(&mut self, new_value: Vec<bool>)

pub fn get_interp_ds(&self) -> Pyo3ArrayF64

pub fn set_interp_ds(&mut self, new_value: Vec<f64>)

pub fn get_interp_is(&self) -> Pyo3ArrayF64

pub fn set_interp_is(&mut self, new_value: Vec<f64>)

pub fn get_interp_hs(&self) -> Pyo3ArrayF64

pub fn set_interp_hs(&mut self, new_value: Vec<f64>)

pub fn get_grades(&self) -> Pyo3ArrayF64

pub fn set_grades(&mut self, new_value: Vec<f64>)

pub fn copy(&self) -> Self

pub fn __copy__(&self) -> Self

pub fn __deepcopy__(&self, _memo: &PyDict) -> Self

pub fn from_resource_py( filepath: &PyAny, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a resource file packaged with the fastsim-core crate

Arguments:

filepath: str | pathlib.Path - Filepath, relative to the top of the resources folder (excluding any relevant prefix), from which to read the object

pub fn to_file_py(&self, filepath: &PyAny) -> PyResult<()>

Write (serialize) an object to a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS. Creates a new file if it does not already exist, otherwise truncates the existing file.

Arguments

filepath: str | pathlib.Path - The filepath at which to write the object

pub fn from_file_py(filepath: &PyAny, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS.

Arguments:

filepath: str | pathlib.Path - The filepath from which to read the object

pub fn to_str_py(&self, format: &str) -> PyResult<String>

Write (serialize) an object into a string

Arguments:

format: str - The target format, any of those listed in ACCEPTED_STR_FORMATS

pub fn from_str_py( contents: &str, format: &str, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a string

Arguments:

contents: str - The string containing the object data

format: str - The source format, any of those listed in ACCEPTED_STR_FORMATS

pub fn to_json_py(&self) -> PyResult<String>

Write (serialize) an object to a JSON string

pub fn from_json_py(json_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object to a JSON string

Arguments

json_str: str - JSON-formatted string to deserialize from

pub fn to_yaml_py(&self) -> PyResult<String>

Write (serialize) an object to a YAML string

pub fn from_yaml_py(yaml_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a YAML string

Arguments

yaml_str: str - YAML-formatted string to deserialize from

impl RustCycleCache

pub fn new(cyc: &RustCycle) -> Self

pub fn interp_grade(&self, dist_m: f64) -> f64

Interpolate the single-point grade at the given distance. Assumes that the grade at i applies from sample point (i-1, i]

pub fn interp_elevation(&self, dist_m: f64) -> f64

Interpolate the elevation at the given distance