Struct fastsim_core::pyo3imports::PyRefMut

pub struct PyRefMut<'p, T>
where
    T: PyClass<Frozen = False>,
{ /* private fields */ }

A wrapper type for a mutably borrowed value from aPyCell<T>.

See the module-level documentation for more information.

Implementations

impl<'p, T> PyRefMut<'p, T>
where T: PyClass<Frozen = False>,

pub fn py(&self) -> Python<'p>

Returns a Python token that is bound to the lifetime of the PyRefMut.

impl<'p, T, U> PyRefMut<'p, T>
where U: PyClass<Frozen = False>, T: PyClass<BaseType = U, Frozen = False>,

pub fn into_super(self) -> PyRefMut<'p, U>

Gets a PyRef<T::BaseType>.

See PyRef::into_super for more.