Trait fastsim_core::pyo3imports::PyTryFrom

pub trait PyTryFrom<'v>: Sized + PyNativeType {
    // Required methods
    fn try_from<V>(value: V) -> Result<&'v Self, PyDowncastError<'v>>
       where V: Into<&'v PyAny>;
    fn try_from_exact<V>(value: V) -> Result<&'v Self, PyDowncastError<'v>>
       where V: Into<&'v PyAny>;
    unsafe fn try_from_unchecked<V>(value: V) -> &'v Self
       where V: Into<&'v PyAny>;
}

Trait implemented by Python object types that allow a checked downcast. If T implements PyTryFrom, we can convert &PyAny to &T.

This trait is similar to std::convert::TryFrom

Required Methods

fn try_from<V>(value: V) -> Result<&'v Self, PyDowncastError<'v>>
where V: Into<&'v PyAny>,

Cast from a concrete Python object type to PyObject.

fn try_from_exact<V>(value: V) -> Result<&'v Self, PyDowncastError<'v>>
where V: Into<&'v PyAny>,

Cast from a concrete Python object type to PyObject. With exact type check.

unsafe fn try_from_unchecked<V>(value: V) -> &'v Self
where V: Into<&'v PyAny>,

Cast a PyAny to a specific type of PyObject. The caller must have already verified the reference is for this type.

Safety

Callers must ensure that the type is valid or risk type confusion.

Object Safety

This trait is not object safe.