Struct fastsim_core::simdrive::RustSimDrive

pub struct RustSimDrive {

    pub hev_sim_count: usize,
    pub veh: RustVehicle,
    pub cyc: RustCycle,
    pub cyc0: RustCycle,
    pub sim_params: RustSimDriveParams,
    pub props: RustPhysicalProperties,
    pub i: usize,
    pub cur_max_fs_kw_out: Array1<f64>,
    pub fc_trans_lim_kw: Array1<f64>,
    pub cur_max_fc_kw_out: Array1<f64>,
    pub ess_cap_lim_dischg_kw: Array1<f64>,
    pub cur_ess_max_kw_out: Array1<f64>,
    pub cur_max_avail_elec_kw: Array1<f64>,
    pub ess_cap_lim_chg_kw: Array1<f64>,
    pub cur_max_ess_chg_kw: Array1<f64>,
    pub cur_max_elec_kw: Array1<f64>,
    pub mc_elec_in_lim_kw: Array1<f64>,
    pub mc_transi_lim_kw: Array1<f64>,
    pub cur_max_mc_kw_out: Array1<f64>,
    pub ess_lim_mc_regen_perc_kw: Array1<f64>,
    pub cur_max_mech_mc_kw_in: Array1<f64>,
    pub cur_max_trans_kw_out: Array1<f64>,
    pub cyc_trac_kw_req: Array1<f64>,
    pub cur_max_trac_kw: Array1<f64>,
    pub spare_trac_kw: Array1<f64>,
    pub cyc_whl_rad_per_sec: Array1<f64>,
    pub cyc_tire_inertia_kw: Array1<f64>,
    pub cyc_whl_kw_req: Array1<f64>,
    pub regen_contrl_lim_kw_perc: Array1<f64>,
    pub cyc_regen_brake_kw: Array1<f64>,
    pub cyc_fric_brake_kw: Array1<f64>,
    pub cyc_trans_kw_out_req: Array1<f64>,
    pub cyc_met: Array1<bool>,
    pub trans_kw_out_ach: Array1<f64>,
    pub trans_kw_in_ach: Array1<f64>,
    pub cur_soc_target: Array1<f64>,
    pub min_mc_kw_2help_fc: Array1<f64>,
    pub mc_mech_kw_out_ach: Array1<f64>,
    pub mc_elec_kw_in_ach: Array1<f64>,
    pub aux_in_kw: Array1<f64>,
    pub impose_coast: Array1<bool>,
    pub roadway_chg_kw_out_ach: Array1<f64>,
    pub min_ess_kw_2help_fc: Array1<f64>,
    pub ess_kw_out_ach: Array1<f64>,
    pub fc_kw_out_ach: Array1<f64>,
    pub fc_kw_out_ach_pct: Array1<f64>,
    pub fc_kw_in_ach: Array1<f64>,
    pub fs_kw_out_ach: Array1<f64>,
    pub fs_kwh_out_ach: Array1<f64>,
    pub ess_cur_kwh: Array1<f64>,
    pub soc: Array1<f64>,
    pub regen_buff_soc: Array1<f64>,
    pub ess_regen_buff_dischg_kw: Array1<f64>,
    pub max_ess_regen_buff_chg_kw: Array1<f64>,
    pub ess_accel_buff_chg_kw: Array1<f64>,
    pub accel_buff_soc: Array1<f64>,
    pub max_ess_accell_buff_dischg_kw: Array1<f64>,
    pub ess_accel_regen_dischg_kw: Array1<f64>,
    pub mc_elec_in_kw_for_max_fc_eff: Array1<f64>,
    pub elec_kw_req_4ae: Array1<f64>,
    pub can_pwr_all_elec: Array1<bool>,
    pub desired_ess_kw_out_for_ae: Array1<f64>,
    pub ess_ae_kw_out: Array1<f64>,
    pub er_ae_kw_out: Array1<f64>,
    pub ess_desired_kw_4fc_eff: Array1<f64>,
    pub ess_kw_if_fc_req: Array1<f64>,
    pub cur_max_mc_elec_kw_in: Array1<f64>,
    pub fc_kw_gap_fr_eff: Array1<f64>,
    pub er_kw_if_fc_req: Array1<f64>,
    pub mc_elec_kw_in_if_fc_req: Array1<f64>,
    pub mc_kw_if_fc_req: Array1<f64>,
    pub fc_forced_on: Array1<bool>,
    pub fc_forced_state: Array1<u32>,
    pub mc_mech_kw_4forced_fc: Array1<f64>,
    pub fc_time_on: Array1<f64>,
    pub prev_fc_time_on: Array1<f64>,
    pub mps_ach: Array1<f64>,
    pub mph_ach: Array1<f64>,
    pub dist_m: Array1<f64>,
    pub dist_mi: Array1<f64>,
    pub high_acc_fc_on_tag: Array1<bool>,
    pub reached_buff: Array1<bool>,
    pub max_trac_mps: Array1<f64>,
    pub add_kwh: Array1<f64>,
    pub dod_cycs: Array1<f64>,
    pub ess_perc_dead: Array1<f64>,
    pub drag_kw: Array1<f64>,
    pub ess_loss_kw: Array1<f64>,
    pub accel_kw: Array1<f64>,
    pub ascent_kw: Array1<f64>,
    pub rr_kw: Array1<f64>,
    pub cur_max_roadway_chg_kw: Array1<f64>,
    pub trace_miss_iters: Array1<u32>,
    pub newton_iters: Array1<u32>,
    pub fuel_kj: f64,
    pub ess_dischg_kj: f64,
    pub energy_audit_error: f64,
    pub mpgge: f64,
    pub roadway_chg_kj: f64,
    pub battery_kwh_per_mi: f64,
    pub electric_kwh_per_mi: f64,
    pub ess2fuel_kwh: f64,
    pub drag_kj: f64,
    pub ascent_kj: f64,
    pub rr_kj: f64,
    pub brake_kj: f64,
    pub trans_kj: f64,
    pub mc_kj: f64,
    pub ess_eff_kj: f64,
    pub aux_kj: f64,
    pub fc_kj: f64,
    pub net_kj: f64,
    pub ke_kj: f64,
    pub trace_miss: bool,
    pub trace_miss_dist_frac: f64,
    pub trace_miss_time_frac: f64,
    pub trace_miss_speed_mps: f64,
    pub orphaned: bool,
    pub coast_delay_index: Array1<i32>,
    pub idm_target_speed_m_per_s: Array1<f64>,
    pub cyc0_cache: RustCycleCache,
    /* private fields */

}

Fields
hev_sim_count: usizeveh: RustVehiclecyc: RustCyclecyc0: RustCyclesim_params: RustSimDriveParamsprops: RustPhysicalPropertiesi: usizecur_max_fs_kw_out: Array1<f64>

Current maximum fuel storage output power, considering veh.fs_max_kw and transient limit, as determined by achieved fuel storage power output and veh.fs_secs_to_peak_pwr

fc_trans_lim_kw: Array1<f64>

Transient fuel converter output power limit, as determined by achieved fuel converter power output, veh.fc_max_kw, and veh.fs_secs_to_peak_pwr

cur_max_fc_kw_out: Array1<f64>

Current maximum fuel converter output power, considering veh.fc_max_kw and transient limit fc_trans_lim_kw

ess_cap_lim_dischg_kw: Array1<f64>

ESS discharging power limit, considering remaining ESS energy and ESS efficiency

cur_ess_max_kw_out: Array1<f64>

Current maximum ESS output power, considering ess_cap_lim_dischg_kw and veh.ess_max_kw

cur_max_avail_elec_kw: Array1<f64>

Current maximum electrical power that can go toward propulsion, cur_max_elec_kw limited by the maximum theoretical motor input power veh.mc_max_elec_in_kw

ess_cap_lim_chg_kw: Array1<f64>

ESS charging power limit, considering unused energy capacity and ESS efficiency

cur_max_ess_chg_kw: Array1<f64>

ESS charging power limit, considering ess_cap_lim_chg_kw and veh.ess_max_kw

cur_max_elec_kw: Array1<f64>

Current maximum electrical power that can go toward propulsion: if FCEV, equal to cur_max_fc_kw_out + cur_max_roadway_chg_kw + cur_ess_max_kw_out - aux_in_kw, otherwise equal to cur_max_roadway_chg_kw + cur_ess_max_kw_out - aux_in_kw

mc_elec_in_lim_kw: Array1<f64>mc_transi_lim_kw: Array1<f64>

Transient electric motor output power limit, as determined by achieved motor mechanical power output, veh.mc_max_kw, and veh.ms_secs_to_peak_pwr

cur_max_mc_kw_out: Array1<f64>ess_lim_mc_regen_perc_kw: Array1<f64>cur_max_mech_mc_kw_in: Array1<f64>

ESS limit on electricity regeneration, considering veh.mc_max_kw, or cur_max_ess_chg_kw and motor efficiency

cur_max_trans_kw_out: Array1<f64>cyc_trac_kw_req: Array1<f64>

Required tractive power to meet cycle, equal to drag_kw + accel_kw + ascent_kw

cur_max_trac_kw: Array1<f64>spare_trac_kw: Array1<f64>cyc_whl_rad_per_sec: Array1<f64>cyc_tire_inertia_kw: Array1<f64>

Power to change wheel rotational speed, calculated with veh.wheel_inertia_kg_m2 and veh.num_wheels

cyc_whl_kw_req: Array1<f64>

Required power to wheels to meet cycle, equal to cyc_trac_kw_req + rr_kw + cyc_tire_inertia_kw

regen_contrl_lim_kw_perc: Array1<f64>cyc_regen_brake_kw: Array1<f64>cyc_fric_brake_kw: Array1<f64>

Power lost to friction braking, only nonzero when cyc_whl_kw_req is negative and regenerative braking cannot provide enough braking,

cyc_trans_kw_out_req: Array1<f64>

Required transmission output power to meet cycle, equal to cyc_whl_kw_req + cyc_fric_brake_kw

cyc_met: Array1<bool>

true if cyc_trans_kw_out_req <= cur_max_trans_kw_out

trans_kw_out_ach: Array1<f64>

Achieved transmission output power, either cyc_trans_kw_out_req if cycle is met, or cur_max_trans_kw_out if it is not

trans_kw_in_ach: Array1<f64>

Achieved transmission input power, accounting for veh.trans_eff

cur_soc_target: Array1<f64>min_mc_kw_2help_fc: Array1<f64>mc_mech_kw_out_ach: Array1<f64>

Achieved electric motor mechanical output power to transmission

mc_elec_kw_in_ach: Array1<f64>

Achieved electric motor electrical input power, accounting for electric motor efficiency

aux_in_kw: Array1<f64>

Auxiliary power load, optionally overridden with an input array, or if aux loads are forced to go through alternator (when veh.no_elec_aux is true) equal to veh.aux_kw / veh.alt_eff otherwise equal to veh.aux_kw

impose_coast: Array1<bool>roadway_chg_kw_out_ach: Array1<f64>min_ess_kw_2help_fc: Array1<f64>ess_kw_out_ach: Array1<f64>fc_kw_out_ach: Array1<f64>fc_kw_out_ach_pct: Array1<f64>fc_kw_in_ach: Array1<f64>fs_kw_out_ach: Array1<f64>fs_kwh_out_ach: Array1<f64>ess_cur_kwh: Array1<f64>soc: Array1<f64>

Current ESS state of charge, multiply by veh.ess_max_kwh to calculate remaining ESS energy

regen_buff_soc: Array1<f64>ess_regen_buff_dischg_kw: Array1<f64>max_ess_regen_buff_chg_kw: Array1<f64>ess_accel_buff_chg_kw: Array1<f64>accel_buff_soc: Array1<f64>max_ess_accell_buff_dischg_kw: Array1<f64>ess_accel_regen_dischg_kw: Array1<f64>mc_elec_in_kw_for_max_fc_eff: Array1<f64>elec_kw_req_4ae: Array1<f64>

Electrical power requirement for all-electric operation, only applicable if vehicle has electrified powertrain, equal to aux_in_kw + trans_kw_in_ach / motor efficiency

can_pwr_all_elec: Array1<bool>desired_ess_kw_out_for_ae: Array1<f64>ess_ae_kw_out: Array1<f64>er_ae_kw_out: Array1<f64>

Charging power received from electric roadway (er), if enabled, for all electric (ae) operation.

ess_desired_kw_4fc_eff: Array1<f64>ess_kw_if_fc_req: Array1<f64>cur_max_mc_elec_kw_in: Array1<f64>fc_kw_gap_fr_eff: Array1<f64>er_kw_if_fc_req: Array1<f64>mc_elec_kw_in_if_fc_req: Array1<f64>mc_kw_if_fc_req: Array1<f64>fc_forced_on: Array1<bool>fc_forced_state: Array1<u32>mc_mech_kw_4forced_fc: Array1<f64>

Power the motor (mc) must provide if the engine (fc) is being forced on. If the engine just turned on and triggers a regen event, it’ll be negative.

fc_time_on: Array1<f64>prev_fc_time_on: Array1<f64>mps_ach: Array1<f64>mph_ach: Array1<f64>dist_m: Array1<f64>dist_mi: Array1<f64>high_acc_fc_on_tag: Array1<bool>reached_buff: Array1<bool>max_trac_mps: Array1<f64>add_kwh: Array1<f64>dod_cycs: Array1<f64>ess_perc_dead: Array1<f64>drag_kw: Array1<f64>

Power lost to aerodynamic drag according to the drag equation, 1/2 * rho * Cd * A * v_avg³ / 1000

ess_loss_kw: Array1<f64>accel_kw: Array1<f64>

Power to accelerate, veh.veh_kg * (v_current² - v_prev²)/2 / dt / 1000

ascent_kw: Array1<f64>

Power expended to ascend a grade, sin(atan(grade)) * props.a_grav_mps2 * veh.veh_kg * v_avg / 1000

rr_kw: Array1<f64>

Power lost to rolling resistance, normal force * veh.wheel_rr_coef * v_avg / 1000, with normal force calculated as cos(atan(grade)) * veh.veh_kg * props.a_grav_mps2

cur_max_roadway_chg_kw: Array1<f64>trace_miss_iters: Array1<u32>newton_iters: Array1<u32>fuel_kj: f64ess_dischg_kj: f64energy_audit_error: f64mpgge: f64roadway_chg_kj: f64battery_kwh_per_mi: f64electric_kwh_per_mi: f64ess2fuel_kwh: f64drag_kj: f64ascent_kj: f64rr_kj: f64brake_kj: f64trans_kj: f64mc_kj: f64ess_eff_kj: f64aux_kj: f64fc_kj: f64net_kj: f64ke_kj: f64trace_miss: bool

true when the vehicle misses the prescribed speed trace

trace_miss_dist_frac: f64

fractional difference between achieved cumulative distance and prescribed cumulative distance

trace_miss_time_frac: f64

fractional difference between achieved time when trace miss is and prescribed cumulative distance

trace_miss_speed_mps: f64

Maximum speed by which vehicle’s speed falls behind prescribed speed trace

orphaned: boolcoast_delay_index: Array1<i32>idm_target_speed_m_per_s: Array1<f64>cyc0_cache: RustCycleCache
Implementations

impl RustSimDrive

pub fn gap_to_lead_vehicle_m(&self) -> Array1<f64>

Provides the gap-with lead vehicle from start to finish

pub fn activate_eco_cruise_rust( &mut self, by_microtrip: bool, extend_fraction: f64, blend_factor: f64, min_target_speed_m_per_s: f64, ) -> Result<()>

Sets the intelligent driver model parameters for an eco-cruise driving trajectory. This is a convenience method instead of setting the sim_params.idm* parameters yourself.

by_microtrip: bool, if True, target speed is set by microtrip, else by cycle

extend_fraction: float, the fraction of time to extend the cycle to allow for catch-up of the following vehicle

blend_factor: float, a value between 0 and 1; only used of by_microtrip is True, blends between microtrip average speed and microtrip average speed when moving. Must be between 0 and 1 inclusive

min_target_speed_m_per_s: float, the minimum speed allowed by the eco-cruise algorithm Mutates the current SimDrive object for eco-cruise.

pub fn next_speed_by_idm( &mut self, i: usize, a_m_per_s2: f64, b_m_per_s2: f64, dt_headway_s: f64, s0_m: f64, v_desired_m_per_s: f64, delta: f64, ) -> f64

Calculate the next speed by the Intelligent Driver Model

i: int, the index

a_m_per_s2: number, max acceleration (m/s2)

b_m_per_s2: number, max deceleration (m/s2)

dt_headway_s: number, the headway between us and the lead vehicle in seconds

s0_m: number, the initial gap between us and the lead vehicle in meters

v_desired_m_per_s: number, the desired speed in (m/s)

delta: number, a shape parameter; typical value is 4.0 RETURN: number, the next speed (m/s) REFERENCE: Treiber, Martin and Kesting, Arne. 2013. “Chapter 11: Car-Following Models Based on Driving Strategies”. Traffic Flow Dynamics: Data, Models and Simulation. Springer-Verlag. Springer, Berlin, Heidelberg. DOI: https://doi.org/10.1007/978-3-642-32460-4.

pub fn set_speed_for_target_gap_using_idm(&mut self, i: usize)

Set gap

i: non-negative integer, the step index RETURN: None EFFECTS:

sets the next speed (m/s) EQUATION: parameters:

v_desired: the desired speed (m/s)

delta: number, typical value is 4.0

a: max acceleration, (m/s2)

b: max deceleration, (m/s2) s = d_lead - d dv/dt = a * (1 - (v/v_desired)**delta - (s_desired(v,v-v_lead)/s)**2) s_desired(v, dv) = s0 + max(0, vdt_headway + (v * dv)/(2.0 * sqrt(ab))) REFERENCE: Treiber, Martin and Kesting, Arne. 2013. “Chapter 11: Car-Following Models Based on Driving Strategies”. Traffic Flow Dynamics: Data, Models and Simulation. Springer-Verlag. Springer, Berlin, Heidelberg. DOI: https://doi.org/10.1007/978-3-642-32460-4

pub fn set_speed_for_target_gap(&mut self, i: usize)

i: non-negative integer, the step index RETURN: None EFFECTS:

sets the next speed (m/s)

pub fn estimate_grade_for_step(&self, i: usize) -> f64

Provides a quick estimate for grade based only on the distance traveled at the start of the current step. If the grade is constant over the step, this is both quick and accurate. NOTE: If not allowing coasting (i.e., sim_params.coast_allow == False) and not allowing IDM/following (i.e., self.sim_params.idm_allow == False) then returns self.cyc.grade[i]

pub fn lookup_grade_for_step(&self, i: usize, mps_ach: Option<f64>) -> f64

For situations where cyc can deviate from cyc0, this method looks up and accurately interpolates what the average grade over the step should be. If mps_ach is not None, the mps_ach value is used to predict the distance traveled over the step. NOTE: If not allowing coasting (i.e., sim_params.coast_allow == False) and not allowing IDM/following (i.e., self.sim_params.idm_allow == False) then returns self.cyc.grade[i]

pub fn set_time_dilation(&mut self, i: usize) -> Result<()>

pub fn set_coast_speed(&mut self, i: usize) -> Result<()>

Placeholder for method to impose coasting. Might be good to include logic for deciding when to coast. Solve for the next-step speed that will yield a zero roadload

impl RustSimDrive

pub fn new(cyc: RustCycle, veh: RustVehicle) -> Self

pub fn len(&self) -> usize

Return length of time arrays

pub fn is_empty(&self) -> bool

Is cycle empty?

pub fn init_arrays(&mut self)

pub fn sim_drive( &mut self, init_soc: Option<f64>, aux_in_kw_override: Option<Array1<f64>>, ) -> Result<()>

Initialize and run sim_drive_walk as appropriate for vehicle attribute vehPtType. Arguments

init_soc: initial SOC for electrified vehicles.
aux_in_kw: aux_in_kw override. Array of same length as cyc.time_s.
Default of None causes veh.aux_kw to be used.

pub fn sim_drive_accel( &mut self, init_soc: Option<f64>, aux_in_kw_override: Option<Array1<f64>>, ) -> Result<()>

pub fn walk( &mut self, init_soc: f64, aux_in_kw_override: Option<Array1<f64>>, ) -> Result<()>

Receives second-by-second cycle information, vehicle properties, and an initial state of charge and runs sim_drive_step to perform a backward facing powertrain simulation. Method sim_drive runs this iteratively to achieve correct SOC initial and final conditions, as needed.

Arguments

init_soc: initial battery state-of-charge (SOC) for electrified vehicles aux_in_kw: (Optional) aux_in_kw override. Array of same length as cyc.time_s. None causes veh.aux_kw to be used.

pub fn init_for_step( &mut self, init_soc: f64, aux_in_kw_override: Option<Array1<f64>>, ) -> Result<()>

This is a specialty method which should be called prior to using sim_drive_step in a loop. Arguments

init_soc: initial battery state-of-charge (SOC) for electrified vehicles aux_in_kw: aux_in_kw override. Array of same length as cyc.time_s.
Default of None causes veh.aux_kw to be used.

pub fn step(&mut self) -> Result<()>

Step through 1 time step.

pub fn solve_step(&mut self, i: usize) -> Result<()>

Perform all the calculations to solve 1 time step.

pub fn set_misc_calcs(&mut self, i: usize) -> Result<()>

Sets misc. calculations at time step ‘i’ Arguments:

i: index of time step

pub fn set_comp_lims(&mut self, i: usize) -> Result<()>

Sets component limits for time step ‘i’ Arguments

i: index of time step initSoc: initial SOC for electrified vehicles

pub fn set_power_calcs(&mut self, i: usize) -> Result<()>

Calculate power requirements to meet cycle and determine if cycle can be met. Arguments

i: index of time step

pub fn set_ach_speed(&mut self, i: usize) -> Result<()>

pub fn set_hybrid_cont_calcs(&mut self, i: usize) -> Result<()>

Hybrid control calculations. Arguments

i: index of time step

pub fn set_fc_forced_state_rust(&mut self, i: usize) -> Result<()>

Calculate control variables related to engine on/off state Arguments

i: index of time step

pub fn set_hybrid_cont_decisions(&mut self, i: usize) -> Result<()>

Hybrid control decisions. Arguments

i: index of time step

pub fn set_fc_power(&mut self, i: usize) -> Result<()>

Sets power consumption values for the current time step. Arguments

i: index of time step

pub fn set_post_scalars(&mut self) -> Result<()>

Sets scalar variables that can be calculated after a cycle is run. This includes mpgge, various energy metrics, and others

impl RustSimDrive

pub fn __new__(cyc: RustCycle, veh: RustVehicle) -> Self

method for instantiating SimDriveRust

pub fn __getnewargs__(&self)

pub fn gap_to_lead_vehicle_m_py(&self) -> Result<Vec<f64>>

Provides the gap-with lead vehicle from start to finish

pub fn sim_drive_py( &mut self, init_soc: Option<f64>, aux_in_kw_override: Option<Vec<f64>>, ) -> Result<()>

Initialize and run sim_drive_walk as appropriate for vehicle attribute vehPtType. Arguments

init_soc: initial SOC for electrified vehicles. aux_in_kw: aux_in_kw override. Array of same length as cyc.time_s. Default of None causes veh.aux_kw to be used.

pub fn sim_drive_walk( &mut self, init_soc: f64, aux_in_kw_override: Option<Vec<f64>>, ) -> Result<()>

Receives second-by-second cycle information, vehicle properties, and an initial state of charge and runs sim_drive_step to perform a backward facing powertrain simulation. Method ‘sim_drive’ runs this iteratively to achieve correct SOC initial and final conditions, as needed.

Arguments

init_soc (optional): initial battery state-of-charge (SOC) for electrified vehicles aux_in_kw: aux_in_kw override. Array of same length as cyc.time_s. None causes veh.aux_kw to be used.

pub fn activate_eco_cruise( &mut self, by_microtrip: Option<bool>, extend_fraction: Option<f64>, blend_factor: Option<f64>, min_target_speed_m_per_s: Option<f64>, ) -> Result<()>

Sets the intelligent driver model parameters for an eco-cruise driving trajectory. This is a convenience method instead of setting the sim_params.idm* parameters yourself.

by_microtrip: bool, if True, target speed is set by microtrip, else by cycle

extend_fraction: float, the fraction of time to extend the cycle to allow for catch-up of the following vehicle

blend_factor: float, a value between 0 and 1; only used of by_microtrip is True, blends between microtrip average speed and microtrip average speed when moving. Must be between 0 and 1 inclusive

pub fn init_for_step_py( &mut self, init_soc: f64, aux_in_kw_override: Option<Vec<f64>>, ) -> Result<()>

This is a specialty method which should be called prior to using sim_drive_step in a loop. Arguments

init_soc: initial battery state-of-charge (SOC) for electrified vehicles aux_in_kw: aux_in_kw override. Array of same length as cyc.time_s. Default of None causes veh.aux_kw to be used.

pub fn sim_drive_step(&mut self) -> Result<()>

Step through 1 time step.

pub fn solve_step_py(&mut self, i: usize) -> Result<()>

Perform all the calculations to solve 1 time step.

pub fn set_misc_calcs_py(&mut self, i: usize) -> Result<()>

Sets misc. calculations at time step ‘i’ Arguments:

i: index of time step

pub fn set_comp_lims_py(&mut self, i: usize) -> Result<()>

pub fn set_power_calcs_py(&mut self, i: usize) -> Result<()>

Calculate power requirements to meet cycle and determine if cycle can be met. Arguments

i: index of time step

pub fn set_ach_speed_py(&mut self, i: usize) -> Result<()>

pub fn set_hybrid_cont_calcs_py(&mut self, i: usize) -> Result<()>

Hybrid control calculations. Arguments

i: index of time step

pub fn set_fc_forced_state_py(&mut self, i: usize) -> Result<()>

Calculate control variables related to engine on/off state Arguments

i: index of time step _py extension is needed to avoid name collision with getter/setter methods

pub fn set_hybrid_cont_decisions_py(&mut self, i: usize) -> Result<()>

Hybrid control decisions. Arguments

i: index of time step

pub fn set_fc_power_py(&mut self, i: usize) -> Result<()>

Sets power consumption values for the current time step. Arguments

i: index of time step

pub fn set_time_dilation_py(&mut self, i: usize) -> Result<()>

Sets the time dilation for the current step. Arguments

i: index of time step

pub fn set_post_scalars_py(&mut self) -> Result<()>

Sets scalar variables that can be calculated after a cycle is run. This includes mpgge, various energy metrics, and others

pub fn len_py(&self) -> usize

pub fn is_empty_py(&self) -> bool

pub fn get_fs_cumu_mj_out_ach(&self) -> Pyo3ArrayF64

pub fn get_fc_cumu_mj_out_ach(&self) -> Pyo3ArrayF64

pub fn get_hev_sim_count(&self) -> usize

pub fn set_hev_sim_count(&mut self, new_value: usize) -> Result<()>

pub fn get_veh(&mut self) -> RustVehicle

pub fn set_veh(&mut self, new_value: RustVehicle) -> Result<()>

pub fn get_cyc(&mut self) -> RustCycle

pub fn set_cyc(&mut self, new_value: RustCycle) -> Result<()>

pub fn get_cyc0(&mut self) -> RustCycle

pub fn set_cyc0(&mut self, new_value: RustCycle) -> Result<()>

pub fn get_sim_params(&mut self) -> RustSimDriveParams

pub fn set_sim_params(&mut self, new_value: RustSimDriveParams) -> Result<()>

pub fn get_props(&mut self) -> RustPhysicalProperties

pub fn set_props(&mut self, new_value: RustPhysicalProperties) -> Result<()>

pub fn get_i(&self) -> usize

pub fn set_i(&mut self, new_value: usize) -> Result<()>

pub fn get_cur_max_fs_kw_out(&self) -> Pyo3ArrayF64

pub fn set_cur_max_fs_kw_out(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_trans_lim_kw(&self) -> Pyo3ArrayF64

pub fn set_fc_trans_lim_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_fc_kw_out(&self) -> Pyo3ArrayF64

pub fn set_cur_max_fc_kw_out(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_cap_lim_dischg_kw(&self) -> Pyo3ArrayF64

pub fn set_ess_cap_lim_dischg_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_ess_max_kw_out(&self) -> Pyo3ArrayF64

pub fn set_cur_ess_max_kw_out(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_avail_elec_kw(&self) -> Pyo3ArrayF64

pub fn set_cur_max_avail_elec_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_cap_lim_chg_kw(&self) -> Pyo3ArrayF64

pub fn set_ess_cap_lim_chg_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_ess_chg_kw(&self) -> Pyo3ArrayF64

pub fn set_cur_max_ess_chg_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_elec_kw(&self) -> Pyo3ArrayF64

pub fn set_cur_max_elec_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_elec_in_lim_kw(&self) -> Pyo3ArrayF64

pub fn set_mc_elec_in_lim_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_transi_lim_kw(&self) -> Pyo3ArrayF64

pub fn set_mc_transi_lim_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_mc_kw_out(&self) -> Pyo3ArrayF64

pub fn set_cur_max_mc_kw_out(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_lim_mc_regen_perc_kw(&self) -> Pyo3ArrayF64

pub fn set_ess_lim_mc_regen_perc_kw( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_cur_max_mech_mc_kw_in(&self) -> Pyo3ArrayF64

pub fn set_cur_max_mech_mc_kw_in(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_trans_kw_out(&self) -> Pyo3ArrayF64

pub fn set_cur_max_trans_kw_out(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cyc_trac_kw_req(&self) -> Pyo3ArrayF64

pub fn set_cyc_trac_kw_req(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_trac_kw(&self) -> Pyo3ArrayF64

pub fn set_cur_max_trac_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_spare_trac_kw(&self) -> Pyo3ArrayF64

pub fn set_spare_trac_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cyc_whl_rad_per_sec(&self) -> Pyo3ArrayF64

pub fn set_cyc_whl_rad_per_sec(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cyc_tire_inertia_kw(&self) -> Pyo3ArrayF64

pub fn set_cyc_tire_inertia_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cyc_whl_kw_req(&self) -> Pyo3ArrayF64

pub fn set_cyc_whl_kw_req(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_regen_contrl_lim_kw_perc(&self) -> Pyo3ArrayF64

pub fn set_regen_contrl_lim_kw_perc( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_cyc_regen_brake_kw(&self) -> Pyo3ArrayF64

pub fn set_cyc_regen_brake_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cyc_fric_brake_kw(&self) -> Pyo3ArrayF64

pub fn set_cyc_fric_brake_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cyc_trans_kw_out_req(&self) -> Pyo3ArrayF64

pub fn set_cyc_trans_kw_out_req(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cyc_met(&self) -> Pyo3ArrayBool

pub fn set_cyc_met(&mut self, new_value: Vec<bool>) -> Result<()>

pub fn get_trans_kw_out_ach(&self) -> Pyo3ArrayF64

pub fn set_trans_kw_out_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_trans_kw_in_ach(&self) -> Pyo3ArrayF64

pub fn set_trans_kw_in_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_soc_target(&self) -> Pyo3ArrayF64

pub fn set_cur_soc_target(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_min_mc_kw_2help_fc(&self) -> Pyo3ArrayF64

pub fn set_min_mc_kw_2help_fc(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_mech_kw_out_ach(&self) -> Pyo3ArrayF64

pub fn set_mc_mech_kw_out_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_elec_kw_in_ach(&self) -> Pyo3ArrayF64

pub fn set_mc_elec_kw_in_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_aux_in_kw(&self) -> Pyo3ArrayF64

pub fn set_aux_in_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_impose_coast(&self) -> Pyo3ArrayBool

pub fn set_impose_coast(&mut self, new_value: Vec<bool>) -> Result<()>

pub fn get_roadway_chg_kw_out_ach(&self) -> Pyo3ArrayF64

pub fn set_roadway_chg_kw_out_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_min_ess_kw_2help_fc(&self) -> Pyo3ArrayF64

pub fn set_min_ess_kw_2help_fc(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_kw_out_ach(&self) -> Pyo3ArrayF64

pub fn set_ess_kw_out_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_kw_out_ach(&self) -> Pyo3ArrayF64

pub fn set_fc_kw_out_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_kw_out_ach_pct(&self) -> Pyo3ArrayF64

pub fn set_fc_kw_out_ach_pct(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_kw_in_ach(&self) -> Pyo3ArrayF64

pub fn set_fc_kw_in_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fs_kw_out_ach(&self) -> Pyo3ArrayF64

pub fn set_fs_kw_out_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fs_kwh_out_ach(&self) -> Pyo3ArrayF64

pub fn set_fs_kwh_out_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_cur_kwh(&self) -> Pyo3ArrayF64

pub fn set_ess_cur_kwh(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_soc(&self) -> Pyo3ArrayF64

pub fn set_soc(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_regen_buff_soc(&self) -> Pyo3ArrayF64

pub fn set_regen_buff_soc(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_regen_buff_dischg_kw(&self) -> Pyo3ArrayF64

pub fn set_ess_regen_buff_dischg_kw( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_max_ess_regen_buff_chg_kw(&self) -> Pyo3ArrayF64

pub fn set_max_ess_regen_buff_chg_kw( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_ess_accel_buff_chg_kw(&self) -> Pyo3ArrayF64

pub fn set_ess_accel_buff_chg_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_accel_buff_soc(&self) -> Pyo3ArrayF64

pub fn set_accel_buff_soc(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_max_ess_accell_buff_dischg_kw(&self) -> Pyo3ArrayF64

pub fn set_max_ess_accell_buff_dischg_kw( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_ess_accel_regen_dischg_kw(&self) -> Pyo3ArrayF64

pub fn set_ess_accel_regen_dischg_kw( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_mc_elec_in_kw_for_max_fc_eff(&self) -> Pyo3ArrayF64

pub fn set_mc_elec_in_kw_for_max_fc_eff( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_elec_kw_req_4ae(&self) -> Pyo3ArrayF64

pub fn set_elec_kw_req_4ae(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_can_pwr_all_elec(&self) -> Pyo3ArrayBool

pub fn set_can_pwr_all_elec(&mut self, new_value: Vec<bool>) -> Result<()>

pub fn get_desired_ess_kw_out_for_ae(&self) -> Pyo3ArrayF64

pub fn set_desired_ess_kw_out_for_ae( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_ess_ae_kw_out(&self) -> Pyo3ArrayF64

pub fn set_ess_ae_kw_out(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_er_ae_kw_out(&self) -> Pyo3ArrayF64

pub fn set_er_ae_kw_out(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_desired_kw_4fc_eff(&self) -> Pyo3ArrayF64

pub fn set_ess_desired_kw_4fc_eff(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_kw_if_fc_req(&self) -> Pyo3ArrayF64

pub fn set_ess_kw_if_fc_req(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_mc_elec_kw_in(&self) -> Pyo3ArrayF64

pub fn set_cur_max_mc_elec_kw_in(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_kw_gap_fr_eff(&self) -> Pyo3ArrayF64

pub fn set_fc_kw_gap_fr_eff(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_er_kw_if_fc_req(&self) -> Pyo3ArrayF64

pub fn set_er_kw_if_fc_req(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_elec_kw_in_if_fc_req(&self) -> Pyo3ArrayF64

pub fn set_mc_elec_kw_in_if_fc_req(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_kw_if_fc_req(&self) -> Pyo3ArrayF64

pub fn set_mc_kw_if_fc_req(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_forced_on(&self) -> Pyo3ArrayBool

pub fn set_fc_forced_on(&mut self, new_value: Vec<bool>) -> Result<()>

pub fn get_fc_forced_state(&self) -> Pyo3ArrayU32

pub fn set_fc_forced_state(&mut self, new_value: Vec<u32>) -> Result<()>

pub fn get_mc_mech_kw_4forced_fc(&self) -> Pyo3ArrayF64

pub fn set_mc_mech_kw_4forced_fc(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_time_on(&self) -> Pyo3ArrayF64

pub fn set_fc_time_on(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_prev_fc_time_on(&self) -> Pyo3ArrayF64

pub fn set_prev_fc_time_on(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mps_ach(&self) -> Pyo3ArrayF64

pub fn set_mps_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mph_ach(&self) -> Pyo3ArrayF64

pub fn set_mph_ach(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_dist_m(&self) -> Pyo3ArrayF64

pub fn set_dist_m(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_dist_mi(&self) -> Pyo3ArrayF64

pub fn set_dist_mi(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_high_acc_fc_on_tag(&self) -> Pyo3ArrayBool

pub fn set_high_acc_fc_on_tag(&mut self, new_value: Vec<bool>) -> Result<()>

pub fn get_reached_buff(&self) -> Pyo3ArrayBool

pub fn set_reached_buff(&mut self, new_value: Vec<bool>) -> Result<()>

pub fn get_max_trac_mps(&self) -> Pyo3ArrayF64

pub fn set_max_trac_mps(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_add_kwh(&self) -> Pyo3ArrayF64

pub fn set_add_kwh(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_dod_cycs(&self) -> Pyo3ArrayF64

pub fn set_dod_cycs(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_perc_dead(&self) -> Pyo3ArrayF64

pub fn set_ess_perc_dead(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_drag_kw(&self) -> Pyo3ArrayF64

pub fn set_drag_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ess_loss_kw(&self) -> Pyo3ArrayF64

pub fn set_ess_loss_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_accel_kw(&self) -> Pyo3ArrayF64

pub fn set_accel_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_ascent_kw(&self) -> Pyo3ArrayF64

pub fn set_ascent_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_rr_kw(&self) -> Pyo3ArrayF64

pub fn set_rr_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_cur_max_roadway_chg_kw(&self) -> Pyo3ArrayF64

pub fn set_cur_max_roadway_chg_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_trace_miss_iters(&self) -> Pyo3ArrayU32

pub fn set_trace_miss_iters(&mut self, new_value: Vec<u32>) -> Result<()>

pub fn get_newton_iters(&self) -> Pyo3ArrayU32

pub fn set_newton_iters(&mut self, new_value: Vec<u32>) -> Result<()>

pub fn get_fuel_kj(&self) -> f64

pub fn set_fuel_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_dischg_kj(&self) -> f64

pub fn set_ess_dischg_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_energy_audit_error(&self) -> f64

pub fn set_energy_audit_error(&mut self, new_value: f64) -> Result<()>

pub fn get_mpgge(&self) -> f64

pub fn set_mpgge(&mut self, new_value: f64) -> Result<()>

pub fn get_roadway_chg_kj(&self) -> f64

pub fn set_roadway_chg_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_battery_kwh_per_mi(&self) -> f64

pub fn set_battery_kwh_per_mi(&mut self, new_value: f64) -> Result<()>

pub fn get_electric_kwh_per_mi(&self) -> f64

pub fn set_electric_kwh_per_mi(&mut self, new_value: f64) -> Result<()>

pub fn get_ess2fuel_kwh(&self) -> f64

pub fn set_ess2fuel_kwh(&mut self, new_value: f64) -> Result<()>

pub fn get_drag_kj(&self) -> f64

pub fn set_drag_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_ascent_kj(&self) -> f64

pub fn set_ascent_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_rr_kj(&self) -> f64

pub fn set_rr_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_brake_kj(&self) -> f64

pub fn set_brake_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_trans_kj(&self) -> f64

pub fn set_trans_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_kj(&self) -> f64

pub fn set_mc_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_eff_kj(&self) -> f64

pub fn set_ess_eff_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_aux_kj(&self) -> f64

pub fn set_aux_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_kj(&self) -> f64

pub fn set_fc_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_net_kj(&self) -> f64

pub fn set_net_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_ke_kj(&self) -> f64

pub fn set_ke_kj(&mut self, new_value: f64) -> Result<()>

pub fn get_trace_miss(&self) -> bool

pub fn set_trace_miss(&mut self, new_value: bool) -> Result<()>

pub fn get_trace_miss_dist_frac(&self) -> f64

pub fn set_trace_miss_dist_frac(&mut self, new_value: f64) -> Result<()>

pub fn get_trace_miss_time_frac(&self) -> f64

pub fn set_trace_miss_time_frac(&mut self, new_value: f64) -> Result<()>

pub fn get_trace_miss_speed_mps(&self) -> f64

pub fn set_trace_miss_speed_mps(&mut self, new_value: f64) -> Result<()>

pub fn get_orphaned(&self) -> bool

pub fn reset_orphaned(&mut self)

Reset the orphaned flag to false.

pub fn get_coast_delay_index(&self) -> Pyo3ArrayI32

pub fn set_coast_delay_index(&mut self, new_value: Vec<i32>) -> Result<()>

pub fn get_idm_target_speed_m_per_s(&self) -> Pyo3ArrayF64

pub fn set_idm_target_speed_m_per_s( &mut self, new_value: Vec<f64>, ) -> Result<()>

pub fn get_cyc0_cache(&self) -> RustCycleCache

pub fn set_cyc0_cache(&mut self, new_value: RustCycleCache) -> Result<()>

pub fn copy(&self) -> Self

pub fn __copy__(&self) -> Self

pub fn __deepcopy__(&self, _memo: &PyDict) -> Self

pub fn from_resource_py( filepath: &PyAny, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a resource file packaged with the fastsim-core crate

Arguments:

filepath: str | pathlib.Path - Filepath, relative to the top of the resources folder (excluding any relevant prefix), from which to read the object

pub fn to_file_py(&self, filepath: &PyAny) -> PyResult<()>

Write (serialize) an object to a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS. Creates a new file if it does not already exist, otherwise truncates the existing file.

Arguments

filepath: str | pathlib.Path - The filepath at which to write the object

pub fn from_file_py(filepath: &PyAny, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS.

Arguments:

filepath: str | pathlib.Path - The filepath from which to read the object

pub fn to_str_py(&self, format: &str) -> PyResult<String>

Write (serialize) an object into a string

Arguments:

format: str - The target format, any of those listed in ACCEPTED_STR_FORMATS

pub fn from_str_py( contents: &str, format: &str, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a string

Arguments:

contents: str - The string containing the object data

format: str - The source format, any of those listed in ACCEPTED_STR_FORMATS

pub fn to_json_py(&self) -> PyResult<String>

Write (serialize) an object to a JSON string

pub fn from_json_py(json_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object to a JSON string

Arguments

json_str: str - JSON-formatted string to deserialize from

pub fn to_yaml_py(&self) -> PyResult<String>

Write (serialize) an object to a YAML string

pub fn from_yaml_py(yaml_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a YAML string

Arguments

yaml_str: str - YAML-formatted string to deserialize from