Module fastsim_core::simdrivelabel

Module containing classes and methods for calculating label fuel economy.

Structs

LabelFe

LabelFePHEV

Label fuel economy values for a PHEV vehicle

PHEVCycleCalc

Label fuel economy calculations for a specific cycle of a PHEV vehicle

Functions

get_label_fe

get_label_fe_phev

get_label_fe_phev_py

pyo3 version of get_label_fe_phev

get_label_fe_py

pyo3 version of get_label_fe

get_net_accel

get_net_accel_py

pyo3 version of get_net_accel

make_accel_trace

make_accel_trace_py

pyo3 version of make_accel_trace