Struct fastsim_core::thermal::SimDriveHot

pub struct SimDriveHot {
    pub sd: RustSimDrive,
    pub vehthrm: VehicleThermal,
    pub state: ThermalState,
    pub history: ThermalStateHistoryVec,
    pub hvac_model_history: HVACModelHistoryVec,
    /* private fields */
}

Fields
sd: RustSimDrivevehthrm: VehicleThermalstate: ThermalStatehistory: ThermalStateHistoryVechvac_model_history: HVACModelHistoryVec
Implementations

impl SimDriveHot

pub fn __new__( cyc: RustCycle, veh: RustVehicle, vehthrm: VehicleThermal, init_state: Option<ThermalState>, amb_te_deg_c: Option<Vec<f64>>, ) -> Self

method for instantiating SimDriveHot

pub fn gap_to_lead_vehicle_m_py(&self) -> Result<Vec<f64>>

Provides the gap-with lead vehicle from start to finish

pub fn sim_drive_py( &mut self, init_soc: Option<f64>, aux_in_kw_override: Option<Vec<f64>>, ) -> Result<()>

Initialize and run sim_drive_walk as appropriate for vehicle attribute vehPtType. Arguments

init_soc: initial SOC for electrified vehicles. aux_in_kw: aux_in_kw override. Array of same length as cyc.time_s. Default of None causes veh.aux_kw to be used.

pub fn sim_drive_walk( &mut self, init_soc: f64, aux_in_kw_override: Option<Vec<f64>>, )

Receives second-by-second cycle information, vehicle properties, and an initial state of charge and runs sim_drive_step to perform a backward facing powertrain simulation. Method ‘sim_drive’ runs this iteratively to achieve correct SOC initial and final conditions, as needed.

Arguments

init_soc (optional): initial battery state-of-charge (SOC) for electrified vehicles aux_in_kw: aux_in_kw override. Array of same length as cyc.time_s. None causes veh.aux_kw to be used.

pub fn init_for_step_py( &mut self, init_soc: f64, aux_in_kw_override: Option<Vec<f64>>, )

This is a specialty method which should be called prior to using sim_drive_step in a loop. Arguments

init_soc: initial battery state-of-charge (SOC) for electrified vehicles aux_in_kw: aux_in_kw override. Array of same length as cyc.time_s. Default of None causes veh.aux_kw to be used.

pub fn sim_drive_step(&mut self) -> Result<()>

Step through 1 time step.

pub fn solve_step_py(&mut self, i: usize) -> Result<()>

Perform all the calculations to solve 1 time step.

pub fn set_misc_calcs_py(&mut self, i: usize)

Sets misc. calculations at time step ‘i’ Arguments:

i: index of time step

pub fn set_comp_lims_py(&mut self, i: usize) -> Result<()>

pub fn set_power_calcs_py(&mut self, i: usize) -> Result<()>

Calculate power requirements to meet cycle and determine if cycle can be met. Arguments

i: index of time step

pub fn set_ach_speed_py(&mut self, i: usize) -> Result<()>

pub fn set_hybrid_cont_calcs_py(&mut self, i: usize) -> Result<()>

Hybrid control calculations. Arguments

i: index of time step

pub fn set_fc_forced_state_py(&mut self, i: usize) -> Result<()>

Calculate control variables related to engine on/off state Arguments

i: index of time step _py extension is needed to avoid name collision with getter/setter methods

pub fn set_hybrid_cont_decisions_py(&mut self, i: usize) -> Result<()>

Hybrid control decisions. Arguments

i: index of time step

pub fn set_fc_power_py(&mut self, i: usize) -> Result<()>

Sets power consumption values for the current time step. Arguments

i: index of time step

pub fn set_time_dilation_py(&mut self, i: usize) -> Result<()>

Sets the time dilation for the current step. Arguments

i: index of time step

pub fn set_post_scalars_py(&mut self) -> Result<()>

Sets scalar variables that can be calculated after a cycle is run. This includes mpgge, various energy metrics, and others

pub fn get_sd(&mut self) -> RustSimDrive

pub fn set_sd(&mut self, new_value: RustSimDrive) -> Result<()>

pub fn get_vehthrm(&mut self) -> VehicleThermal

pub fn set_vehthrm(&mut self, new_value: VehicleThermal) -> Result<()>

pub fn get_state(&mut self) -> ThermalState

pub fn set_state(&mut self, new_value: ThermalState) -> Result<()>

pub fn get_history(&self) -> ThermalStateHistoryVec

pub fn set_history(&mut self, new_value: ThermalStateHistoryVec) -> Result<()>

pub fn get_hvac_model_history(&self) -> HVACModelHistoryVec

pub fn set_hvac_model_history( &mut self, new_value: HVACModelHistoryVec, ) -> Result<()>

pub fn copy(&self) -> Self

pub fn __copy__(&self) -> Self

pub fn __deepcopy__(&self, _memo: &PyDict) -> Self

pub fn from_resource_py( filepath: &PyAny, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a resource file packaged with the fastsim-core crate

Arguments:

filepath: str | pathlib.Path - Filepath, relative to the top of the resources folder (excluding any relevant prefix), from which to read the object

pub fn to_file_py(&self, filepath: &PyAny) -> PyResult<()>

Write (serialize) an object to a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS. Creates a new file if it does not already exist, otherwise truncates the existing file.

Arguments

filepath: str | pathlib.Path - The filepath at which to write the object

pub fn from_file_py(filepath: &PyAny, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS.

Arguments:

filepath: str | pathlib.Path - The filepath from which to read the object

pub fn to_str_py(&self, format: &str) -> PyResult<String>

Write (serialize) an object into a string

Arguments:

format: str - The target format, any of those listed in ACCEPTED_STR_FORMATS

pub fn from_str_py( contents: &str, format: &str, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a string

Arguments:

contents: str - The string containing the object data

format: str - The source format, any of those listed in ACCEPTED_STR_FORMATS

pub fn to_json_py(&self) -> PyResult<String>

Write (serialize) an object to a JSON string

pub fn from_json_py(json_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object to a JSON string

Arguments

json_str: str - JSON-formatted string to deserialize from

pub fn to_yaml_py(&self) -> PyResult<String>

Write (serialize) an object to a YAML string

pub fn from_yaml_py(yaml_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a YAML string

Arguments

yaml_str: str - YAML-formatted string to deserialize from

impl SimDriveHot

pub fn new( cyc: RustCycle, veh: RustVehicle, vehthrm: VehicleThermal, init_state: Option<ThermalState>, amb_te_deg_c: Option<Array1<f64>>, ) -> Self

pub fn gap_to_lead_vehicle_m(&self) -> Array1<f64>

pub fn sim_drive( &mut self, init_soc: Option<f64>, aux_in_kw_override: Option<Array1<f64>>, ) -> Result<()>

pub fn walk(&mut self, init_soc: f64, aux_in_kw_override: Option<Array1<f64>>)

pub fn init_for_step( &mut self, init_soc: f64, aux_in_kw_override: Option<Array1<f64>>, )

pub fn set_speed_for_target_gap_using_idm(&mut self, i: usize)

pub fn set_speed_for_target_gap(&mut self, i: usize)

pub fn step(&mut self) -> Result<()>

pub fn solve_step(&mut self, i: usize) -> Result<()>

pub fn set_thermal_calcs(&mut self, i: usize)

pub fn set_fc_thermal_calcs(&mut self, i: usize)

Solve fuel converter thermal behavior assuming convection parameters of sphere.

pub fn set_cab_thermal_calcs(&mut self, i: usize)

Solve cabin thermal behavior.

pub fn set_exhport_thermal_calcs(&mut self, i: usize)

Solve exhport thermal behavior.

pub fn thermal_soak_walk(&mut self)

pub fn set_cat_thermal_calcs(&mut self, i: usize)

Solve catalyst thermal behavior.

pub fn set_misc_calcs(&mut self, i: usize)

pub fn set_comp_lims(&mut self, i: usize) -> Result<()>

pub fn set_power_calcs(&mut self, i: usize) -> Result<()>

pub fn set_ach_speed(&mut self, i: usize) -> Result<()>

pub fn set_hybrid_cont_calcs(&mut self, i: usize) -> Result<()>

pub fn set_fc_forced_state_rust(&mut self, i: usize) -> Result<()>

pub fn set_hybrid_cont_decisions(&mut self, i: usize) -> Result<()>

pub fn set_fc_power(&mut self, i: usize) -> Result<()>

pub fn set_time_dilation(&mut self, i: usize) -> Result<()>

pub fn set_post_scalars(&mut self) -> Result<()>