Struct fastsim_core::vehicle::RustVehicle

pub struct RustVehicle {

    pub doc: Option<String>,
    pub props: RustPhysicalProperties,
    pub scenario_name: String,
    pub selection: u32,
    pub veh_year: u32,
    pub veh_pt_type: String,
    pub drag_coef: f64,
    pub drag_coef_doc: Option<String>,
    pub frontal_area_m2: f64,
    pub frontal_area_m2_doc: Option<String>,
    pub glider_kg: f64,
    pub glider_kg_doc: Option<String>,
    pub veh_cg_m: f64,
    pub veh_cg_m_doc: Option<String>,
    pub drive_axle_weight_frac: f64,
    pub drive_axle_weight_frac_doc: Option<String>,
    pub wheel_base_m: f64,
    pub wheel_base_m_doc: Option<String>,
    pub cargo_kg: f64,
    pub cargo_kg_doc: Option<String>,
    pub veh_override_kg: Option<f64>,
    pub veh_override_kg_doc: Option<String>,
    pub comp_mass_multiplier: f64,
    pub comp_mass_multiplier_doc: Option<String>,
    pub fs_max_kw: f64,
    pub fs_max_kw_doc: Option<String>,
    pub fs_secs_to_peak_pwr: f64,
    pub fs_secs_to_peak_pwr_doc: Option<String>,
    pub fs_kwh: f64,
    pub fs_kwh_doc: Option<String>,
    pub fs_kwh_per_kg: f64,
    pub fs_kwh_per_kg_doc: Option<String>,
    pub fc_max_kw: f64,
    pub fc_max_kw_doc: Option<String>,
    pub fc_pwr_out_perc: Array1<f64>,
    pub fc_pwr_out_perc_doc: Option<String>,
    pub fc_eff_map: Array1<f64>,
    pub fc_eff_map_doc: Option<String>,
    pub fc_eff_type: String,
    pub fc_eff_type_doc: Option<String>,
    pub fc_sec_to_peak_pwr: f64,
    pub fc_sec_to_peak_pwr_doc: Option<String>,
    pub fc_base_kg: f64,
    pub fc_base_kg_doc: Option<String>,
    pub fc_kw_per_kg: f64,
    pub fc_kw_per_kg_doc: Option<String>,
    pub min_fc_time_on: f64,
    pub min_fc_time_on_doc: Option<String>,
    pub idle_fc_kw: f64,
    pub idle_fc_kw_doc: Option<String>,
    pub mc_max_kw: f64,
    pub mc_max_kw_doc: Option<String>,
    pub mc_pwr_out_perc: Array1<f64>,
    pub mc_pwr_out_perc_doc: Option<String>,
    pub mc_eff_map: Array1<f64>,
    pub mc_eff_map_doc: Option<String>,
    pub mc_sec_to_peak_pwr: f64,
    pub mc_sec_to_peak_pwr_doc: Option<String>,
    pub mc_pe_kg_per_kw: f64,
    pub mc_pe_kg_per_kw_doc: Option<String>,
    pub mc_pe_base_kg: f64,
    pub mc_pe_base_kg_doc: Option<String>,
    pub ess_max_kw: f64,
    pub ess_max_kw_doc: Option<String>,
    pub ess_max_kwh: f64,
    pub ess_max_kwh_doc: Option<String>,
    pub ess_kg_per_kwh: f64,
    pub ess_kg_per_kwh_doc: Option<String>,
    pub ess_base_kg: f64,
    pub ess_base_kg_doc: Option<String>,
    pub ess_round_trip_eff: f64,
    pub ess_round_trip_eff_doc: Option<String>,
    pub ess_life_coef_a: f64,
    pub ess_life_coef_a_doc: Option<String>,
    pub ess_life_coef_b: f64,
    pub ess_life_coef_b_doc: Option<String>,
    pub min_soc: f64,
    pub min_soc_doc: Option<String>,
    pub max_soc: f64,
    pub max_soc_doc: Option<String>,
    pub ess_dischg_to_fc_max_eff_perc: f64,
    pub ess_dischg_to_fc_max_eff_perc_doc: Option<String>,
    pub ess_chg_to_fc_max_eff_perc: f64,
    pub ess_chg_to_fc_max_eff_perc_doc: Option<String>,
    pub wheel_inertia_kg_m2: f64,
    pub wheel_inertia_kg_m2_doc: Option<String>,
    pub num_wheels: f64,
    pub num_wheels_doc: Option<String>,
    pub wheel_rr_coef: f64,
    pub wheel_rr_coef_doc: Option<String>,
    pub wheel_radius_m: f64,
    pub wheel_radius_m_doc: Option<String>,
    pub wheel_coef_of_fric: f64,
    pub wheel_coef_of_fric_doc: Option<String>,
    pub max_accel_buffer_mph: f64,
    pub max_accel_buffer_mph_doc: Option<String>,
    pub max_accel_buffer_perc_of_useable_soc: f64,
    pub max_accel_buffer_perc_of_useable_soc_doc: Option<String>,
    pub perc_high_acc_buf: f64,
    pub perc_high_acc_buf_doc: Option<String>,
    pub mph_fc_on: f64,
    pub mph_fc_on_doc: Option<String>,
    pub kw_demand_fc_on: f64,
    pub kw_demand_fc_on_doc: Option<String>,
    pub max_regen: f64,
    pub max_regen_doc: Option<String>,
    pub stop_start: bool,
    pub stop_start_doc: Option<String>,
    pub force_aux_on_fc: bool,
    pub force_aux_on_fc_doc: Option<String>,
    pub alt_eff: f64,
    pub alt_eff_doc: Option<String>,
    pub chg_eff: f64,
    pub chg_eff_doc: Option<String>,
    pub aux_kw: f64,
    pub aux_kw_doc: Option<String>,
    pub trans_kg: f64,
    pub trans_kg_doc: Option<String>,
    pub trans_eff: f64,
    pub trans_eff_doc: Option<String>,
    pub ess_to_fuel_ok_error: f64,
    pub ess_to_fuel_ok_error_doc: Option<String>,
    pub fc_peak_eff_override: Option<f64>,
    pub fc_peak_eff_override_doc: Option<String>,
    pub mc_peak_eff_override: Option<f64>,
    pub mc_peak_eff_override_doc: Option<String>,
    /* private fields */

}

Struct containing vehicle attributes

Python Examples

import fastsim

## Load drive cycle by name
cyc_py = fastsim.cycle.Cycle.from_file("udds")
cyc_rust = cyc_py.to_rust()

Fields
doc: Option<String>

Vehicle level documentation – e.g. info about calibration/validation of vehicle and/or links to reports or other long-form documentation.

props: RustPhysicalProperties

Physical properties, see RustPhysicalProperties

scenario_name: String

Vehicle name

selection: u32

Vehicle database ID

veh_year: u32

Vehicle year

veh_pt_type: String

Vehicle powertrain type, one of [CONV, HEV, PHEV, BEV]

drag_coef: f64

Aerodynamic drag coefficient

drag_coef_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

frontal_area_m2: f64

Frontal area, $m^2$

frontal_area_m2_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

glider_kg: f64

Vehicle mass excluding cargo, passengers, and powertrain components, $kg$

glider_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

veh_cg_m: f64

Vehicle center of mass height, $m$ NOTE: positive for FWD, negative for RWD, AWD, 4WD

veh_cg_m_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

drive_axle_weight_frac: f64

Fraction of weight on the drive axle while stopped

drive_axle_weight_frac_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

wheel_base_m: f64

Wheelbase, $m$

wheel_base_m_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

cargo_kg: f64

Cargo mass including passengers, $kg$

cargo_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

veh_override_kg: Option<f64>

Total vehicle mass, overrides mass calculation, $kg$

veh_override_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

comp_mass_multiplier: f64

Component mass multiplier for vehicle mass calculation

comp_mass_multiplier_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fs_max_kw: f64

Fuel storage max power output, $kW$

fs_max_kw_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fs_secs_to_peak_pwr: f64

Fuel storage time to peak power, $s$

fs_secs_to_peak_pwr_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fs_kwh: f64

Fuel storage energy capacity, $kWh$

fs_kwh_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fs_kwh_per_kg: f64

Fuel specific energy, $\frac{kWh}{kg}$

fs_kwh_per_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fc_max_kw: f64

Fuel converter peak continuous power, $kW$

fc_max_kw_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fc_pwr_out_perc: Array1<f64>

Fuel converter output power percentage map, x values of fc_eff_map

fc_pwr_out_perc_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fc_eff_map: Array1<f64>

Fuel converter efficiency map

fc_eff_map_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fc_eff_type: String

Fuel converter efficiency type, one of [SI, ATKINSON, DIESEL, H2FC, HD_DIESEL] Used for calculating fc_eff_map, and other calculations if H2FC

fc_eff_type_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fc_sec_to_peak_pwr: f64

Fuel converter time to peak power, $s$

fc_sec_to_peak_pwr_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fc_base_kg: f64

Fuel converter base mass, $kg$

fc_base_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fc_kw_per_kg: f64

Fuel converter specific power (power-to-weight ratio), $\frac{kW}{kg}$

fc_kw_per_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

min_fc_time_on: f64

Minimum time fuel converter must be on before shutoff (for HEV, PHEV)

min_fc_time_on_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

idle_fc_kw: f64

Fuel converter idle power, $kW$

idle_fc_kw_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

mc_max_kw: f64

Peak continuous electric motor power, $kW$

mc_max_kw_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

mc_pwr_out_perc: Array1<f64>

Electric motor output power percentage map, x values of mc_eff_map

mc_pwr_out_perc_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

mc_eff_map: Array1<f64>

Electric motor efficiency map

mc_eff_map_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

mc_sec_to_peak_pwr: f64

Electric motor time to peak power, $s$

mc_sec_to_peak_pwr_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

mc_pe_kg_per_kw: f64

Motor power electronics mass per power output, $\frac{kg}{kW}$

mc_pe_kg_per_kw_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

mc_pe_base_kg: f64

Motor power electronics base mass, $kg$

mc_pe_base_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_max_kw: f64

Traction battery maximum power output, $kW$

ess_max_kw_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_max_kwh: f64

Traction battery energy capacity, $kWh$

ess_max_kwh_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_kg_per_kwh: f64

Traction battery mass per energy, $\frac{kg}{kWh}$

ess_kg_per_kwh_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_base_kg: f64

Traction battery base mass, $kg$

ess_base_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_round_trip_eff: f64

Traction battery round-trip efficiency

ess_round_trip_eff_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_life_coef_a: f64

Traction battery cycle life coefficient A, see reference

ess_life_coef_a_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_life_coef_b: f64

Traction battery cycle life coefficient B, see reference

ess_life_coef_b_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

min_soc: f64

Traction battery minimum state of charge

min_soc_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

max_soc: f64

Traction battery maximum state of charge

max_soc_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_dischg_to_fc_max_eff_perc: f64

ESS discharge effort toward max FC efficiency

ess_dischg_to_fc_max_eff_perc_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_chg_to_fc_max_eff_perc: f64

ESS charge effort toward max FC efficiency

ess_chg_to_fc_max_eff_perc_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

wheel_inertia_kg_m2: f64

Mass moment of inertia per wheel, $kg \cdot m^2$

wheel_inertia_kg_m2_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

num_wheels: f64

Number of wheels

num_wheels_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

wheel_rr_coef: f64

Rolling resistance coefficient

wheel_rr_coef_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

wheel_radius_m: f64

Wheel radius, $m$

wheel_radius_m_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

wheel_coef_of_fric: f64

Wheel coefficient of friction

wheel_coef_of_fric_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

max_accel_buffer_mph: f64

Speed where the battery reserved for accelerating is zero

max_accel_buffer_mph_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

max_accel_buffer_perc_of_useable_soc: f64

Percent of usable battery energy reserved to help accelerate

max_accel_buffer_perc_of_useable_soc_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

perc_high_acc_buf: f64

Percent SOC buffer for high accessory loads during cycles with long idle time

perc_high_acc_buf_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

mph_fc_on: f64

Speed at which the fuel converter must turn on, $mph$

mph_fc_on_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

kw_demand_fc_on: f64

Power demand above which to require fuel converter on, $kW$

kw_demand_fc_on_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

max_regen: f64

Maximum brake regeneration efficiency

max_regen_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

stop_start: bool

Stop/start micro-HEV flag

stop_start_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

force_aux_on_fc: bool

Force auxiliary power load to come from fuel converter

force_aux_on_fc_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

alt_eff: f64

Alternator efficiency

alt_eff_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

chg_eff: f64

Charger efficiency

chg_eff_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

aux_kw: f64

Auxiliary load power, $kW$

aux_kw_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

trans_kg: f64

Transmission mass, $kg$

trans_kg_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

trans_eff: f64

Transmission efficiency

trans_eff_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

ess_to_fuel_ok_error: f64

Maximum acceptable ratio of change in ESS energy to expended fuel energy (used in hybrid SOC balancing), $\frac{\Delta E_{ESS}}{\Delta E_{fuel}}$

ess_to_fuel_ok_error_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

fc_peak_eff_override: Option<f64>

Fuel converter efficiency peak override, scales entire curve

fc_peak_eff_override_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

mc_peak_eff_override: Option<f64>

Motor efficiency peak override, scales entire curve

mc_peak_eff_override_doc: Option<String>

String for documentation – e.g. info about calibration/validation.

Implementations

impl RustVehicle

pub fn set_veh_mass_py(&mut self)

pub fn get_mc_peak_eff(&self) -> f64

pub fn set_mc_peak_eff_py(&mut self, new_peak: f64)

pub fn get_max_fc_eff_kw(&self) -> f64

pub fn set_fc_peak_eff_py(&mut self, new_peak: f64)

pub fn get_fc_peak_eff(&self) -> f64

pub fn set_derived_py(&mut self)

pub fn to_rust(&self) -> Self

An identify function to allow RustVehicle to be used as a python vehicle and respond to this method Returns a clone of the current object

pub fn list_resources_py(&self) -> Vec<String>

list available vehicle resources

pub fn get_doc(&self) -> Option<String>

pub fn set_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_props(&mut self) -> RustPhysicalProperties

pub fn set_props(&mut self, new_value: RustPhysicalProperties) -> Result<()>

pub fn get_scenario_name(&self) -> String

pub fn set_scenario_name(&mut self, new_value: String) -> Result<()>

pub fn get_selection(&self) -> u32

pub fn set_selection(&mut self, new_value: u32) -> Result<()>

pub fn get_veh_year(&self) -> u32

pub fn set_veh_year(&mut self, new_value: u32) -> Result<()>

pub fn get_veh_pt_type(&self) -> String

pub fn set_veh_pt_type(&mut self, new_value: String) -> Result<()>

pub fn get_drag_coef(&self) -> f64

pub fn set_drag_coef(&mut self, new_value: f64) -> Result<()>

pub fn get_drag_coef_doc(&self) -> Option<String>

pub fn set_drag_coef_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_frontal_area_m2(&self) -> f64

pub fn set_frontal_area_m2(&mut self, new_value: f64) -> Result<()>

pub fn get_frontal_area_m2_doc(&self) -> Option<String>

pub fn set_frontal_area_m2_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_glider_kg(&self) -> f64

pub fn set_glider_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_glider_kg_doc(&self) -> Option<String>

pub fn set_glider_kg_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_veh_cg_m(&self) -> f64

pub fn set_veh_cg_m(&mut self, new_value: f64) -> Result<()>

pub fn get_veh_cg_m_doc(&self) -> Option<String>

pub fn set_veh_cg_m_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_drive_axle_weight_frac(&self) -> f64

pub fn set_drive_axle_weight_frac(&mut self, new_value: f64) -> Result<()>

pub fn get_drive_axle_weight_frac_doc(&self) -> Option<String>

pub fn set_drive_axle_weight_frac_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_wheel_base_m(&self) -> f64

pub fn set_wheel_base_m(&mut self, new_value: f64) -> Result<()>

pub fn get_wheel_base_m_doc(&self) -> Option<String>

pub fn set_wheel_base_m_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_cargo_kg(&self) -> f64

pub fn set_cargo_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_cargo_kg_doc(&self) -> Option<String>

pub fn set_cargo_kg_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_veh_override_kg(&self) -> Option<f64>

pub fn set_veh_override_kg(&mut self, new_value: Option<f64>) -> Result<()>

pub fn get_veh_override_kg_doc(&self) -> Option<String>

pub fn set_veh_override_kg_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_comp_mass_multiplier(&self) -> f64

pub fn set_comp_mass_multiplier(&mut self, new_value: f64) -> Result<()>

pub fn get_comp_mass_multiplier_doc(&self) -> Option<String>

pub fn set_comp_mass_multiplier_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_fs_max_kw(&self) -> f64

pub fn set_fs_max_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_fs_max_kw_doc(&self) -> Option<String>

pub fn set_fs_max_kw_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_fs_secs_to_peak_pwr(&self) -> f64

pub fn set_fs_secs_to_peak_pwr(&mut self, new_value: f64) -> Result<()>

pub fn get_fs_secs_to_peak_pwr_doc(&self) -> Option<String>

pub fn set_fs_secs_to_peak_pwr_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_fs_kwh(&self) -> f64

pub fn set_fs_kwh(&mut self, new_value: f64) -> Result<()>

pub fn get_fs_kwh_doc(&self) -> Option<String>

pub fn set_fs_kwh_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_fs_kwh_per_kg(&self) -> f64

pub fn set_fs_kwh_per_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_fs_kwh_per_kg_doc(&self) -> Option<String>

pub fn set_fs_kwh_per_kg_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_fc_max_kw(&self) -> f64

pub fn set_fc_max_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_max_kw_doc(&self) -> Option<String>

pub fn set_fc_max_kw_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_fc_pwr_out_perc(&self) -> Pyo3ArrayF64

pub fn set_fc_pwr_out_perc(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_pwr_out_perc_doc(&self) -> Option<String>

pub fn set_fc_pwr_out_perc_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_fc_eff_map(&self) -> Pyo3ArrayF64

pub fn set_fc_eff_map(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_eff_map_doc(&self) -> Option<String>

pub fn set_fc_eff_map_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_fc_eff_type(&self) -> String

pub fn set_fc_eff_type(&mut self, new_value: String) -> Result<()>

pub fn get_fc_eff_type_doc(&self) -> Option<String>

pub fn set_fc_eff_type_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_fc_sec_to_peak_pwr(&self) -> f64

pub fn set_fc_sec_to_peak_pwr(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_sec_to_peak_pwr_doc(&self) -> Option<String>

pub fn set_fc_sec_to_peak_pwr_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_fc_base_kg(&self) -> f64

pub fn set_fc_base_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_base_kg_doc(&self) -> Option<String>

pub fn set_fc_base_kg_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_fc_kw_per_kg(&self) -> f64

pub fn set_fc_kw_per_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_kw_per_kg_doc(&self) -> Option<String>

pub fn set_fc_kw_per_kg_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_min_fc_time_on(&self) -> f64

pub fn set_min_fc_time_on(&mut self, new_value: f64) -> Result<()>

pub fn get_min_fc_time_on_doc(&self) -> Option<String>

pub fn set_min_fc_time_on_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_idle_fc_kw(&self) -> f64

pub fn set_idle_fc_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_idle_fc_kw_doc(&self) -> Option<String>

pub fn set_idle_fc_kw_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_mc_max_kw(&self) -> f64

pub fn set_mc_max_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_max_kw_doc(&self) -> Option<String>

pub fn set_mc_max_kw_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_mc_pwr_out_perc(&self) -> Pyo3ArrayF64

pub fn set_mc_pwr_out_perc(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_pwr_out_perc_doc(&self) -> Option<String>

pub fn set_mc_pwr_out_perc_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_mc_eff_map(&self) -> Pyo3ArrayF64

pub fn set_mc_eff_map(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_eff_map_doc(&self) -> Option<String>

pub fn set_mc_eff_map_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_mc_sec_to_peak_pwr(&self) -> f64

pub fn set_mc_sec_to_peak_pwr(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_sec_to_peak_pwr_doc(&self) -> Option<String>

pub fn set_mc_sec_to_peak_pwr_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_mc_pe_kg_per_kw(&self) -> f64

pub fn set_mc_pe_kg_per_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_pe_kg_per_kw_doc(&self) -> Option<String>

pub fn set_mc_pe_kg_per_kw_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_mc_pe_base_kg(&self) -> f64

pub fn set_mc_pe_base_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_pe_base_kg_doc(&self) -> Option<String>

pub fn set_mc_pe_base_kg_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_ess_max_kw(&self) -> f64

pub fn set_ess_max_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_max_kw_doc(&self) -> Option<String>

pub fn set_ess_max_kw_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_ess_max_kwh(&self) -> f64

pub fn set_ess_max_kwh(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_max_kwh_doc(&self) -> Option<String>

pub fn set_ess_max_kwh_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_ess_kg_per_kwh(&self) -> f64

pub fn set_ess_kg_per_kwh(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_kg_per_kwh_doc(&self) -> Option<String>

pub fn set_ess_kg_per_kwh_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_ess_base_kg(&self) -> f64

pub fn set_ess_base_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_base_kg_doc(&self) -> Option<String>

pub fn set_ess_base_kg_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_ess_round_trip_eff(&self) -> f64

pub fn set_ess_round_trip_eff(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_round_trip_eff_doc(&self) -> Option<String>

pub fn set_ess_round_trip_eff_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_ess_life_coef_a(&self) -> f64

pub fn set_ess_life_coef_a(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_life_coef_a_doc(&self) -> Option<String>

pub fn set_ess_life_coef_a_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_ess_life_coef_b(&self) -> f64

pub fn set_ess_life_coef_b(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_life_coef_b_doc(&self) -> Option<String>

pub fn set_ess_life_coef_b_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_min_soc(&self) -> f64

pub fn set_min_soc(&mut self, new_value: f64) -> Result<()>

pub fn get_min_soc_doc(&self) -> Option<String>

pub fn set_min_soc_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_max_soc(&self) -> f64

pub fn set_max_soc(&mut self, new_value: f64) -> Result<()>

pub fn get_max_soc_doc(&self) -> Option<String>

pub fn set_max_soc_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_ess_dischg_to_fc_max_eff_perc(&self) -> f64

pub fn set_ess_dischg_to_fc_max_eff_perc( &mut self, new_value: f64, ) -> Result<()>

pub fn get_ess_dischg_to_fc_max_eff_perc_doc(&self) -> Option<String>

pub fn set_ess_dischg_to_fc_max_eff_perc_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_ess_chg_to_fc_max_eff_perc(&self) -> f64

pub fn set_ess_chg_to_fc_max_eff_perc(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_chg_to_fc_max_eff_perc_doc(&self) -> Option<String>

pub fn set_ess_chg_to_fc_max_eff_perc_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_wheel_inertia_kg_m2(&self) -> f64

pub fn set_wheel_inertia_kg_m2(&mut self, new_value: f64) -> Result<()>

pub fn get_wheel_inertia_kg_m2_doc(&self) -> Option<String>

pub fn set_wheel_inertia_kg_m2_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_num_wheels(&self) -> f64

pub fn set_num_wheels(&mut self, new_value: f64) -> Result<()>

pub fn get_num_wheels_doc(&self) -> Option<String>

pub fn set_num_wheels_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_wheel_rr_coef(&self) -> f64

pub fn set_wheel_rr_coef(&mut self, new_value: f64) -> Result<()>

pub fn get_wheel_rr_coef_doc(&self) -> Option<String>

pub fn set_wheel_rr_coef_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_wheel_radius_m(&self) -> f64

pub fn set_wheel_radius_m(&mut self, new_value: f64) -> Result<()>

pub fn get_wheel_radius_m_doc(&self) -> Option<String>

pub fn set_wheel_radius_m_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_wheel_coef_of_fric(&self) -> f64

pub fn set_wheel_coef_of_fric(&mut self, new_value: f64) -> Result<()>

pub fn get_wheel_coef_of_fric_doc(&self) -> Option<String>

pub fn set_wheel_coef_of_fric_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_max_accel_buffer_mph(&self) -> f64

pub fn set_max_accel_buffer_mph(&mut self, new_value: f64) -> Result<()>

pub fn get_max_accel_buffer_mph_doc(&self) -> Option<String>

pub fn set_max_accel_buffer_mph_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_max_accel_buffer_perc_of_useable_soc(&self) -> f64

pub fn set_max_accel_buffer_perc_of_useable_soc( &mut self, new_value: f64, ) -> Result<()>

pub fn get_max_accel_buffer_perc_of_useable_soc_doc(&self) -> Option<String>

pub fn set_max_accel_buffer_perc_of_useable_soc_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_perc_high_acc_buf(&self) -> f64

pub fn set_perc_high_acc_buf(&mut self, new_value: f64) -> Result<()>

pub fn get_perc_high_acc_buf_doc(&self) -> Option<String>

pub fn set_perc_high_acc_buf_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_mph_fc_on(&self) -> f64

pub fn set_mph_fc_on(&mut self, new_value: f64) -> Result<()>

pub fn get_mph_fc_on_doc(&self) -> Option<String>

pub fn set_mph_fc_on_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_kw_demand_fc_on(&self) -> f64

pub fn set_kw_demand_fc_on(&mut self, new_value: f64) -> Result<()>

pub fn get_kw_demand_fc_on_doc(&self) -> Option<String>

pub fn set_kw_demand_fc_on_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_max_regen(&self) -> f64

pub fn set_max_regen(&mut self, new_value: f64) -> Result<()>

pub fn get_max_regen_doc(&self) -> Option<String>

pub fn set_max_regen_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_stop_start(&self) -> bool

pub fn set_stop_start(&mut self, new_value: bool) -> Result<()>

pub fn get_stop_start_doc(&self) -> Option<String>

pub fn set_stop_start_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_force_aux_on_fc(&self) -> bool

pub fn set_force_aux_on_fc(&mut self, new_value: bool) -> Result<()>

pub fn get_force_aux_on_fc_doc(&self) -> Option<String>

pub fn set_force_aux_on_fc_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_alt_eff(&self) -> f64

pub fn set_alt_eff(&mut self, new_value: f64) -> Result<()>

pub fn get_alt_eff_doc(&self) -> Option<String>

pub fn set_alt_eff_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_chg_eff(&self) -> f64

pub fn set_chg_eff(&mut self, new_value: f64) -> Result<()>

pub fn get_chg_eff_doc(&self) -> Option<String>

pub fn set_chg_eff_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_aux_kw(&self) -> f64

pub fn set_aux_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_aux_kw_doc(&self) -> Option<String>

pub fn set_aux_kw_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_trans_kg(&self) -> f64

pub fn set_trans_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_trans_kg_doc(&self) -> Option<String>

pub fn set_trans_kg_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_trans_eff(&self) -> f64

pub fn set_trans_eff(&mut self, new_value: f64) -> Result<()>

pub fn get_trans_eff_doc(&self) -> Option<String>

pub fn set_trans_eff_doc(&mut self, new_value: Option<String>) -> Result<()>

pub fn get_ess_to_fuel_ok_error(&self) -> f64

pub fn set_ess_to_fuel_ok_error(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_to_fuel_ok_error_doc(&self) -> Option<String>

pub fn set_ess_to_fuel_ok_error_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_small_motor_power_kw(&self) -> f64

pub fn set_small_motor_power_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_large_motor_power_kw(&self) -> f64

pub fn set_large_motor_power_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_perc_out_array(&self) -> Pyo3VecF64

pub fn set_fc_perc_out_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_regen_a(&self) -> f64

pub fn set_regen_a(&mut self, new_value: f64) -> Result<()>

pub fn get_regen_b(&self) -> f64

pub fn set_regen_b(&mut self, new_value: f64) -> Result<()>

pub fn get_charging_on(&self) -> bool

pub fn set_charging_on(&mut self, new_value: bool) -> Result<()>

pub fn get_no_elec_sys(&self) -> bool

pub fn set_no_elec_sys(&mut self, new_value: bool) -> Result<()>

pub fn get_no_elec_aux(&self) -> bool

pub fn set_no_elec_aux(&mut self, new_value: bool) -> Result<()>

pub fn get_max_roadway_chg_kw(&self) -> Pyo3ArrayF64

pub fn set_max_roadway_chg_kw(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_input_kw_out_array(&self) -> Pyo3ArrayF64

pub fn set_input_kw_out_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_kw_out_array(&self) -> Pyo3VecF64

pub fn set_fc_kw_out_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_fc_eff_array(&self) -> Pyo3VecF64

pub fn set_fc_eff_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_modern_max(&self) -> f64

pub fn set_modern_max(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_eff_array(&self) -> Pyo3ArrayF64

pub fn set_mc_eff_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_kw_in_array(&self) -> Pyo3VecF64

pub fn set_mc_kw_in_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_kw_out_array(&self) -> Pyo3VecF64

pub fn set_mc_kw_out_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_mc_max_elec_in_kw(&self) -> f64

pub fn set_mc_max_elec_in_kw(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_full_eff_array(&self) -> Pyo3VecF64

pub fn set_mc_full_eff_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_veh_kg(&self) -> f64

pub fn set_veh_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_max_trac_mps2(&self) -> f64

pub fn set_max_trac_mps2(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_mass_kg(&self) -> f64

pub fn set_ess_mass_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_mass_kg(&self) -> f64

pub fn set_mc_mass_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_mass_kg(&self) -> f64

pub fn set_fc_mass_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_fs_mass_kg(&self) -> f64

pub fn set_fs_mass_kg(&mut self, new_value: f64) -> Result<()>

pub fn get_mc_perc_out_array(&self) -> Pyo3VecF64

pub fn set_mc_perc_out_array(&mut self, new_value: Vec<f64>) -> Result<()>

pub fn get_val_udds_mpgge(&self) -> f64

pub fn set_val_udds_mpgge(&mut self, new_value: f64) -> Result<()>

pub fn get_val_hwy_mpgge(&self) -> f64

pub fn set_val_hwy_mpgge(&mut self, new_value: f64) -> Result<()>

pub fn get_val_comb_mpgge(&self) -> f64

pub fn set_val_comb_mpgge(&mut self, new_value: f64) -> Result<()>

pub fn get_val_udds_kwh_per_mile(&self) -> f64

pub fn set_val_udds_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val_hwy_kwh_per_mile(&self) -> f64

pub fn set_val_hwy_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val_comb_kwh_per_mile(&self) -> f64

pub fn set_val_comb_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val_cd_range_mi(&self) -> f64

pub fn set_val_cd_range_mi(&mut self, new_value: f64) -> Result<()>

pub fn get_val_const65_mph_kwh_per_mile(&self) -> f64

pub fn set_val_const65_mph_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val_const60_mph_kwh_per_mile(&self) -> f64

pub fn set_val_const60_mph_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val_const55_mph_kwh_per_mile(&self) -> f64

pub fn set_val_const55_mph_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val_const45_mph_kwh_per_mile(&self) -> f64

pub fn set_val_const45_mph_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val_unadj_udds_kwh_per_mile(&self) -> f64

pub fn set_val_unadj_udds_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val_unadj_hwy_kwh_per_mile(&self) -> f64

pub fn set_val_unadj_hwy_kwh_per_mile(&mut self, new_value: f64) -> Result<()>

pub fn get_val0_to60_mph(&self) -> f64

pub fn set_val0_to60_mph(&mut self, new_value: f64) -> Result<()>

pub fn get_val_ess_life_miles(&self) -> f64

pub fn set_val_ess_life_miles(&mut self, new_value: f64) -> Result<()>

pub fn get_val_range_miles(&self) -> f64

pub fn set_val_range_miles(&mut self, new_value: f64) -> Result<()>

pub fn get_val_veh_base_cost(&self) -> f64

pub fn set_val_veh_base_cost(&mut self, new_value: f64) -> Result<()>

pub fn get_val_msrp(&self) -> f64

pub fn set_val_msrp(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_peak_eff_override(&self) -> Option<f64>

pub fn set_fc_peak_eff_override(&mut self, new_value: Option<f64>) -> Result<()>

pub fn get_fc_peak_eff_override_doc(&self) -> Option<String>

pub fn set_fc_peak_eff_override_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_mc_peak_eff_override(&self) -> Option<f64>

pub fn set_mc_peak_eff_override(&mut self, new_value: Option<f64>) -> Result<()>

pub fn get_mc_peak_eff_override_doc(&self) -> Option<String>

pub fn set_mc_peak_eff_override_doc( &mut self, new_value: Option<String>, ) -> Result<()>

pub fn get_orphaned(&self) -> bool

pub fn reset_orphaned(&mut self)

Reset the orphaned flag to false.

pub fn get_max_regen_kwh(&self) -> f64

pub fn set_max_regen_kwh(&mut self, new_value: f64) -> Result<()>

pub fn copy(&self) -> Self

pub fn __copy__(&self) -> Self

pub fn __deepcopy__(&self, _memo: &PyDict) -> Self

pub fn from_resource_py( filepath: &PyAny, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a resource file packaged with the fastsim-core crate

Arguments:

filepath: str | pathlib.Path - Filepath, relative to the top of the resources folder (excluding any relevant prefix), from which to read the object

pub fn to_file_py(&self, filepath: &PyAny) -> PyResult<()>

Write (serialize) an object to a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS. Creates a new file if it does not already exist, otherwise truncates the existing file.

Arguments

filepath: str | pathlib.Path - The filepath at which to write the object

pub fn from_file_py(filepath: &PyAny, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS.

Arguments:

filepath: str | pathlib.Path - The filepath from which to read the object

pub fn to_str_py(&self, format: &str) -> PyResult<String>

Write (serialize) an object into a string

Arguments:

format: str - The target format, any of those listed in ACCEPTED_STR_FORMATS

pub fn from_str_py( contents: &str, format: &str, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a string

Arguments:

contents: str - The string containing the object data

format: str - The source format, any of those listed in ACCEPTED_STR_FORMATS

pub fn to_json_py(&self) -> PyResult<String>

Write (serialize) an object to a JSON string

pub fn from_json_py(json_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object to a JSON string

Arguments

json_str: str - JSON-formatted string to deserialize from

pub fn to_yaml_py(&self) -> PyResult<String>

Write (serialize) an object to a YAML string

pub fn from_yaml_py(yaml_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a YAML string

Arguments

yaml_str: str - YAML-formatted string to deserialize from

impl RustVehicle

RustVehicle rust methods

pub fn set_veh_mass(&mut self)

Sets the following parameters:

ess_mass_kg

mc_mass_kg

fc_mass_kg

fs_mass_kg

veh_kg

max_trac_mps2

pub fn mc_peak_eff(&self) -> f64

pub fn max_fc_eff_kw(&self) -> f64

Returns first FC output power at which peak efficiency occurs

pub fn fc_peak_eff(&self) -> f64

pub fn set_mc_peak_eff(&mut self, new_peak: f64)

pub fn set_fc_peak_eff(&mut self, new_peak: f64)

pub fn set_derived(&mut self) -> Result<()>

Sets derived parameters:

no_elec_sys

no_elec_aux

fc_perc_out_array

input_kw_out_array

fc_kw_out_array

fc_eff_array

modern_diff

large_baseline_eff_adj

mc_kw_adj_perc

mc_eff_map

mc_eff_array

mc_perc_out_array

mc_kw_out_array

mc_full_eff_array

mc_kw_in_array

mc_max_elec_in_kw

set_fc_peak_eff()

set_mc_peak_eff()

set_veh_mass()

ess_mass_kg

mc_mass_kg

fc_mass_kg

fs_mass_kg

veh_kg

max_trac_mps2

pub fn mock_vehicle() -> Self

pub fn from_github_or_url<S: AsRef<str>>( vehicle_file_name: S, url: Option<S>, ) -> Result<Self>

Downloads specified vehicle from FASTSim vehicle repo or url and instantiates it into a RustVehicle. Notes in vehicle.doc the origin of the vehicle. Returns vehicle.

Arguments

vehicle_file_name: file name for vehicle to be downloaded, including path from url directory or FASTSim repository (if applicable)

url: url for vehicle repository where vehicle will be downloaded from, if None, assumed to be downloaded from vehicle FASTSim repo

Note: The URL needs to be a URL pointing directly to a file, for example a raw github URL, split up so that the “url” argument is the path to the directory, and the “vehicle_file_name” is the path within the directory to the file.
Note: If downloading from the FASTSim Vehicle Repo, the vehicle_file_name should include the path to the file from the root of the Repo, as listed in the output of the vehicle_utils::fetch_github_list() function.
Note: the url should not include the file name, only the path to the file or a root directory of the file.