Module fastsim_core::vehicle_utils

Module for utility functions that support the vehicle struct.

Structs

GitObjectInfo

ObjectLinks

Functions

abc_to_drag_coeffs

extract_zip

Extract zip archive at filepath to destination directory at dest_dir

fetch_github_list

Returns a list of vehicle file names in the Fastsim Vehicle Repo, or, optionally, a different GitHub repo, in which case the url provided needs to be the url for the file tree within GitHub for the root folder the Rust objects, for example “https://api.github.com/repos/NREL/fastsim-vehicles/contents/public”
Note: for each file, the output will list the vehicle file name, including the path from the root of the repository

get_error_val

list_zip_contents

Given the path to a zip archive, print out the names of the files within that archive